+++ src/pflow/core/prompt_cache_analysis/analyze.py
@@ -22,10 +22,10 @@ from __future__ import annotations
 
 import logging
 from collections.abc import Callable, Iterable, Mapping, Sequence
-from dataclasses import dataclass, field, replace
+from dataclasses import dataclass, replace
 from datetime import datetime, timezone
 from pathlib import Path
-from typing import Any, Final
+from typing import Any
 
 # Imported for the predicted-cache_key contract (Round 4 high-value fix #2):
 # when ``cache.discrepancy`` detection wires up in v1.x, it MUST use the shared
@@ -36,10 +36,9 @@ from typing import Any, Final
 # defers full prediction; the helpers will be consumed when that lands.
 from pflow.core.cache_ttl import parse_cache_ttl
 from pflow.core.diagnostic import Diagnostic, Severity
-from pflow.core.exceptions import CompilationError, MarkdownParseError, SchemaValidationError, WorkflowValidationError
-from pflow.core.llm_capabilities import anthropic_models_at_threshold, get_min_cache_tokens
+from pflow.core.llm_capabilities import get_min_cache_tokens
 from pflow.core.llm_config import get_default_workflow_model
-from pflow.core.llm_providers import detect_provider, normalize_model_name
+from pflow.core.llm_providers import normalize_model_name
 from pflow.core.llm_usage import normalize_litellm_usage_tokens
 from pflow.core.prompt_cache import (  # noqa: F401 — see docstring.
     _CHUNK_ABSENT,
@@ -63,9 +62,22 @@ from .below_min_tokens_detector import (
 from .below_min_tokens_detector import (
     detect as detect_below_min_tokens,
 )
-from .context import _PREDICTION_SKIPPED, AnalysisContext, _latest_memo_for_freshness_check, _normalize_empty
-from .cross_workflow import CrossWorkflowEdge, DynamicBatchInfo, walk_cross_workflow
+from .context import AnalysisContext, _normalize_empty
+from .cross_workflow import CrossWorkflowEdge, walk_cross_workflow
 from .padding_advisor import PaddingCandidate, compute_padding_advisories
+from .stages.discrepancy import (
+    _attach_predicted_cache_keys,
+    _emit_discrepancy_diagnostics,
+    _format_dynamic_batches_note,
+)
+from .stages.summary import (
+    _aggregate_confidence,
+    _build_summary,
+    _filter_trace_dependent_warnings,
+    _format_workflow_run_command,
+    _maybe_append_gemini_note,
+    _trace_coverage_for_rows,
+)
 from .token_estimation import (
     _estimate_ref_tokens,
     estimate_cacheable_tokens,
@@ -88,7 +100,6 @@ from .trace_loading import (
     _diagnostics_from_trace_warnings,
     _edge_child_paths,
     _format_rejection_note,
-    _is_llm_node,
     _resolve_ir_static_model_for_node,
     _resolve_trace_data,
     _resolve_trace_scope,
@@ -99,14 +110,9 @@ from .types import (
     _BLOCK_BELOW_PROVIDER_MIN,
     _BLOCK_PREWARM_IMAGES,
     _BLOCK_RUNTIME_STRIPPED,
-    _PROJECTION_NOT_APPLICABLE,
-    _PROJECTION_UNAVAILABLE,
-    AnalysisSummary,
     CacheAnalysis,
     CacheProjection,
     CacheProjectionComponent,
-    CostDelta,
-    CrossWorkflowFindings,
     CrossWorkflowInputContribution,
     PerCallRow,
     PerNodeThresholdEntry,
@@ -115,7 +121,6 @@ from .types import (
     SuggestedBlock,
     SuggestedBlockChunk,
     TraceExecutionIndex,
-    TraceUnexecutedLLMRow,
     _projection_component,
     _projection_source_confidence,
     _provider_min_state,
@@ -123,7 +128,7 @@ from .types import (
     aggregate_projection,
     invocation_count_for,
 )
-from .warning_catalog import CACHE_WARNING_CATALOG, make_diagnostic
+from .warning_catalog import make_diagnostic
 
 logger = logging.getLogger(__name__)
 
@@ -132,19 +137,6 @@ _SUGGESTED_BLOCK_BELOW_THRESHOLD: str = "below_threshold"
 _SUGGESTED_BLOCK_EVIDENCE_INCOMPLETE: str = "evidence_incomplete"
 _SUGGESTED_BLOCK_INSUFFICIENT_NODES: str = "insufficient_nodes"
 _PARENT_PROSE_PREVIEW_LIMIT = 40
-
-
-_PREDICTION_RECOVERABLE_EXCEPTIONS: Final[tuple[type[Exception], ...]] = (
-    CompilationError,
-    MarkdownParseError,
-    SchemaValidationError,
-    WorkflowValidationError,
-    FileNotFoundError,
-    ValueError,
-    KeyError,
-    RecursionError,
-    OSError,
-)
 
 
 # ---------------------------------------------------------------------------
@@ -429,6 +421,8 @@ def analyze(
     # Stage 0: walker now returns (graph_info, findings). Findings flow into
     # ``warnings`` (single source of truth); renderers categorize at output
     # time by filtering on ``Diagnostic.id``.
+    from .stages.cross_workflow import _build_cross_workflow_findings
+
     cross_findings, cross_diagnostics = _build_cross_workflow_findings(
         cw_result=cw_result,
         notes=notes,
@@ -1123,6 +1117,15 @@ def _row_cross_workflow_candidates_for_edge(
     trace_index: TraceExecutionIndex,
 ) -> tuple[_RowCrossWorkflowCandidate, ...]:
     """Return gated row-level candidates for one cross-workflow edge."""
+    from .stages.cross_workflow import (
+        _append_child_suffix,
+        _cache_ref_is_declared_or_covered,
+        _child_cache_ref_consumers,
+        _estimate_parent_value_tokens,
+        _items_by_name,
+        _parent_prose_for_cache_ref,
+    )
+
     if getattr(edge, "is_batch_alias_root", False) or getattr(edge, "parent_value_expr", None) is None:
         return ()
     child_workflow = str(edge.child_workflow)
@@ -1189,6 +1192,12 @@ def _row_cross_workflow_candidates_for_edge(
 
 def _has_structural_cross_workflow_projection_candidate(cw_result: Any) -> bool:
     """Return True when a no-trace run has rows that trace attribution could fill."""
+    from .stages.cross_workflow import (
+        _cache_ref_is_declared_or_covered,
+        _child_cache_ref_consumers,
+        _items_by_name,
+    )
+
     for edge in getattr(cw_result, "edges", ()) or ():
         if getattr(edge, "is_batch_alias_root", False):
             continue
@@ -4175,20 +4184,6 @@ def _emit_padding_advisories(
     return compute_padding_advisories(candidates)
 
 
-def _is_cache_focused(diag: Diagnostic) -> bool:
-    """Whether a diagnostic belongs to provider prompt-cache analysis.
-
-    The analyzer runs the full validator pipeline, but headline counts and
-    advisory actions remain cache-domain signals. Catalog-IDed cache findings
-    use the cache warning catalog; intentionally un-IDed cache reference errors
-    carry paths under ``cache.*`` or containing ``.prompt_cache``.
-    """
-    if diag.id and diag.id in CACHE_WARNING_CATALOG:
-        return True
-    path = (diag.context or {}).get("path")
-    return isinstance(path, str) and (path.startswith("cache.") or ".prompt_cache" in path)
-
-
 def _run_full_validation(
     workflow_ir: dict[str, Any],
     *,
@@ -4322,2514 +4317,6 @@ def _input_rate(model: str) -> float | None:
 
     pricing = get_model_pricing(model)
     return pricing.input_rate if pricing is not None else None
-
-
-# ---------------------------------------------------------------------------
-# Cross-workflow walking
-# ---------------------------------------------------------------------------
-
-
-def _build_cross_workflow_findings(
-    *,
-    cw_result: Any,
-    notes: list[str],
-    per_call_rows: list[PerCallRow],
-    ctx: AnalysisContext,
-    call_counts_by_node: dict[tuple[str | None, str], int],
-) -> tuple[CrossWorkflowFindings, list[Diagnostic]]:
-    """Run the F1.3 walker and emit rename / prose-mismatch / value-flow diagnostics.
-
-    Stage 0 (Task 159): returns ``(graph_info, findings)``. Diagnostics flow
-    into the analyzer's single ``warnings`` list; the renderers categorize at
-    output time by filtering on ``Diagnostic.id``. The graph_info field carries
-    pure topology (``boundaries_analyzed``) for JSON consumers that need the
-    edge count without inspecting individual findings.
-
-    The walker appends notes to the supplied list when it stops descending
-    a branch (max_depth or cycle) — the analyzer surfaces these via
-    ``CacheAnalysis.notes`` so agents see truncation rather than silent
-    incompleteness.
-    """
-    result = cw_result
-    edges = result.edges
-
-    rename_diags: list[Diagnostic] = []
-    prose_mismatches: list[Diagnostic] = []
-    sub_workflow_cache_candidates: list[_SubWorkflowCacheCandidate] = []
-    for edge in edges:
-        is_rename = bool(edge.is_rename and edge.parent_value_expr is not None)
-        if is_rename:
-            # Evidence-basis principle: the rename warning predicts that
-            # cross-workflow byte-level cache match WILL fail because of
-            # diverging prose labels. That prediction is only meaningful
-            # when (a) the parent value is a stable identifier (not a
-            # batch iteration variable), and (b) at least one side has
-            # ``## Cache`` declared so there's actual state to break.
-            # Without these, the warning fires hypothetically and floods
-            # the agent-facing report with non-actionable noise. See
-            # GH #362 for the empirical case (lyrics-generator: 23
-            # rename warnings, all from batch aliases or non-cache
-            # boundaries — zero actionable).
-            if edge.is_batch_alias_root:
-                continue  # Iteration-variable substitution, not a rename.
-            parent_has_cache = bool(result.cache_items_by_workflow.get(edge.parent_workflow))
-            child_has_cache = bool(result.cache_items_by_workflow.get(edge.child_workflow))
-            if parent_has_cache or child_has_cache:
-                rename_diags.append(
-                    make_diagnostic(
-                        "cache.cross-workflow-rename-detected",
-                        parent_workflow=edge.parent_workflow,
-                        child_workflow=edge.child_workflow,
-                        parent_value_expr=edge.parent_value_expr,
-                        child_input_name=edge.child_input_name,
-                        line_in_parent=edge.line_in_parent,
-                        parent_node_id=edge.parent_node_id,
-                    )
-                )
-
-        if not is_rename:
-            prose_mismatches.extend(_cross_workflow_prose_mismatches(edge, result.cache_items_by_workflow))
-        sub_workflow_cache_candidates.extend(
-            _sub_workflow_cache_candidates_for_edge(edge, result.cache_items_by_workflow, result.irs_by_workflow)
-        )
-
-    rows_by_node_path: dict[tuple[str | None, str], PerCallRow] = {
-        (row.workflow_path, row.node_path): row for row in per_call_rows
-    }
-    sub_workflow_cache_diags = _emit_sub_workflow_cache_findings(
-        sub_workflow_cache_candidates,
-        rows_by_node_path=rows_by_node_path,
-        ctx=ctx,
-        cw_result=cw_result,
-        call_counts_by_node=call_counts_by_node,
-    )
-
-    findings: list[Diagnostic] = [*rename_diags, *prose_mismatches, *sub_workflow_cache_diags]
-    return (CrossWorkflowFindings(boundaries_analyzed=len(edges)), findings)
-
-
-def _cross_workflow_prose_mismatches(
-    edge: Any,
-    cache_items_by_workflow: dict[str, tuple[dict[str, Any], ...]],
-) -> list[Diagnostic]:
-    parent_by_name = _items_by_name(cache_items_by_workflow.get(edge.parent_workflow, ()))
-    child_by_name = _items_by_name(cache_items_by_workflow.get(edge.child_workflow, ()))
-    diagnostics: list[Diagnostic] = []
-    for chunk_name in sorted(parent_by_name.keys() & child_by_name.keys()):
-        parent_prose = str(parent_by_name[chunk_name].get("prose_before", ""))
-        child_prose = str(child_by_name[chunk_name].get("prose_before", ""))
-        if parent_prose == child_prose:
-            continue
-        diagnostics.append(
-            make_diagnostic(
-                "cache.cross-workflow-prose-mismatch",
-                parent_workflow=edge.parent_workflow,
-                child_workflow=edge.child_workflow,
-                chunk_name=chunk_name,
-                parent_prose=parent_prose,
-                child_prose=child_prose,
-            )
-        )
-    return diagnostics
-
-
-@dataclass(frozen=True)
-class _SubWorkflowCacheCandidate:
-    """One child-workflow cache declaration opportunity.
-
-    Sub-workflows do not inherit parent ``## Cache`` blocks, so a parent-side
-    declaration never satisfies the child. The actionable edit is in the
-    receiving workflow when that child has repeated LLM consumers of the
-    incoming input.
-    """
-
-    parent_workflow: str
-    parent_value_expr: str
-    parent_node_id: str
-    line_in_parent: int
-    child_workflow: str
-    child_input_name: str
-    child_count: int
-    child_node_ids: tuple[str, ...]
-    parent_cache_ref: str = ""
-    child_cache_ref: str = ""
-    parent_prose: str = ""
-    parent_prose_origins_differ: bool = False
-    body_refs_by_node: Mapping[str, tuple[str, ...]] = field(default_factory=dict)
-    origin_count: int = 1
-    has_multiple_parent_origins: bool = False
-
-    def __post_init__(self) -> None:
-        if not self.child_cache_ref:
-            object.__setattr__(self, "child_cache_ref", self.child_input_name)
-        if not self.parent_cache_ref:
-            object.__setattr__(self, "parent_cache_ref", self.parent_value_expr)
-
-
-@dataclass(frozen=True)
-class _ChildCacheRefUse:
-    """Actual child prompt refs under a workflow input, grouped by cache entry."""
-
-    child_cache_ref: str
-    consumer_node_ids: tuple[str, ...]
-    body_refs_by_node: Mapping[str, tuple[str, ...]]
-
-
-@dataclass(frozen=True)
-class _GroupedConsumerProjection:
-    """One child LLM node's cumulative cache opportunity within a group.
-
-    ``per_call_prefix_tokens`` is the sum of input tokens this consumer receives
-    on each call. It's what the provider's cache minimum is checked against —
-    pflow emits one or more ``cache_control`` markers per call (Anthropic
-    multi-breakpoint); the terminal marker sees the full per-call prefix bytes.
-    The threshold gate must use this field (per-call prefix tokens), NOT a
-    cohort total (per_call × call_count), which mixes units and silently
-    mis-classifies low-per-call high-call-count cases. Multi-breakpoint
-    placement does not change the per-call denominator.
-
-    ``threshold`` is the provider's per-call minimum for this consumer's model.
-    ``savings_usd`` is cohort-level (already multiplied by call count) — the
-    cost projection is honest about per-workflow-run savings.
-    """
-
-    consumer_node_id: str
-    model: str
-    consumed_inputs: tuple[str, ...]
-    per_call_prefix_tokens: int
-    threshold: int
-    savings_usd: float | None
-    did_not_execute_in_trace: bool
-
-
-@dataclass(frozen=True)
-class _SubWorkflowCacheGroup:
-    """All undeclared cache inputs flowing into one child workflow."""
-
-    child_workflow: str
-    candidates: tuple[_SubWorkflowCacheCandidate, ...]
-
-
-def _sub_workflow_cache_candidates_for_edge(
-    edge: Any,
-    cache_items_by_workflow: dict[str, tuple[dict[str, Any], ...]],
-    irs_by_workflow: dict[str, dict[str, Any]],
-) -> tuple[_SubWorkflowCacheCandidate, ...]:
-    """Return prompt-ref-level candidates for one boundary edge.
-
-    Suppression rules:
-    - ``parent_value_expr is None``: literal or multi-ref string at the
-      boundary; no template to track.
-    - the child already declares the receiving ref, or an ancestor ref, in its
-      own ``## Cache``.
-    - no LLM nodes in the child consume the ref.
-
-    Batch-scoped parent values are still valid here: ``${item.concept}`` varies
-    across parent fanout items, but inside each child invocation the receiving
-    input can be stable context reused by multiple child LLM nodes.
-    """
-    if edge.parent_value_expr is None:
-        return ()
-    child_declared = set(_items_by_name(cache_items_by_workflow.get(edge.child_workflow, ())))
-    parent_items_by_name = _items_by_name(cache_items_by_workflow.get(edge.parent_workflow, ()))
-
-    candidates: list[_SubWorkflowCacheCandidate] = []
-    for use in _child_cache_ref_consumers(irs_by_workflow.get(edge.child_workflow, {}), edge.child_input_name):
-        if _cache_ref_is_declared_or_covered(use.child_cache_ref, child_declared):
-            continue
-        parent_cache_ref = _append_child_suffix(edge.parent_value_expr, edge.child_input_name, use.child_cache_ref)
-        parent_prose = _parent_prose_for_cache_ref(parent_items_by_name, parent_cache_ref)
-        candidates.append(
-            _SubWorkflowCacheCandidate(
-                parent_workflow=edge.parent_workflow,
-                parent_value_expr=edge.parent_value_expr,
-                parent_cache_ref=parent_cache_ref,
-                parent_prose=parent_prose,
-                parent_node_id=edge.parent_node_id,
-                line_in_parent=edge.line_in_parent,
-                child_workflow=edge.child_workflow,
-                child_input_name=edge.child_input_name,
-                child_cache_ref=use.child_cache_ref,
-                child_count=len(use.consumer_node_ids),
-                child_node_ids=use.consumer_node_ids,
-                body_refs_by_node=use.body_refs_by_node,
-            )
-        )
-    return tuple(candidates)
-
-
-def _child_cache_ref_consumers(child_ir: dict[str, Any], child_input_name: str) -> tuple[_ChildCacheRefUse, ...]:
-    """Return prompt refs under ``child_input_name``, grouped by cache entry."""
-    if not isinstance(child_ir, dict):
-        return ()
-    by_ref: dict[str, dict[str, list[str]]] = {}
-    first_seen: list[str] = []
-    for node in child_ir.get("nodes", []) or []:
-        node_id, node_refs = _node_child_cache_ref_body_refs(node, child_input_name)
-        if node_id is None:
-            continue
-        for child_cache_ref, body_refs in node_refs.items():
-            if child_cache_ref not in by_ref:
-                by_ref[child_cache_ref] = {}
-                first_seen.append(child_cache_ref)
-            by_ref[child_cache_ref][node_id] = body_refs
-    uses: list[_ChildCacheRefUse] = []
-    for child_cache_ref in first_seen:
-        refs_by_node = by_ref[child_cache_ref]
-        uses.append(
-            _ChildCacheRefUse(
-                child_cache_ref=child_cache_ref,
-                consumer_node_ids=tuple(refs_by_node),
-                body_refs_by_node={node_id: tuple(refs) for node_id, refs in refs_by_node.items()},
-            )
-        )
-    return tuple(uses)
-
-
-def _node_child_cache_ref_body_refs(
-    node: Any,
-    child_input_name: str,
-) -> tuple[str | None, dict[str, list[str]]]:
-    if not isinstance(node, dict) or node.get("type") != "llm":
-        return None, {}
-    node_id_raw = node.get("id")
-    prompt = node.get("params", {}).get("prompt", "")
-    if not isinstance(node_id_raw, str) or not node_id_raw or not isinstance(prompt, str):
-        return None, {}
-    node_refs: dict[str, list[str]] = {}
-    for ref in classify_prompt_refs(prompt, batch_alias=None, node_inputs=_node_inputs(node)):
-        for operand in ref.operand_paths:
-            if _path_is_equal_or_descendant(operand, child_input_name):
-                node_refs.setdefault(operand, []).append(ref.raw_expr)
-    if child_input_name in node_refs:
-        node_refs = {child_input_name: node_refs[child_input_name]}
-    return node_id_raw, {child_ref: list(dict.fromkeys(refs)) for child_ref, refs in node_refs.items()}
-
-
-def _path_is_equal_or_descendant(path: str, root: str) -> bool:
-    """True when ``path`` is ``root`` or a segment-aware descendant."""
-    if not path or not root:
-        return False
-    return path == root or path.startswith(f"{root}.") or path.startswith(f"{root}[")
-
-
-def _path_is_ancestor_or_equal(ancestor: str, child: str) -> bool:
-    return _path_is_equal_or_descendant(child, ancestor)
-
-
-def _child_suffix(child_input_name: str, child_ref: str) -> str:
-    """Return the suffix of ``child_ref`` after ``child_input_name``."""
-    if child_ref == child_input_name:
-        return ""
-    if child_ref.startswith(f"{child_input_name}."):
-        return child_ref[len(child_input_name) :]
-    if child_ref.startswith(f"{child_input_name}["):
-        return child_ref[len(child_input_name) :]
-    return ""
-
-
-def _append_child_suffix(parent_ref: str, child_input_name: str, child_ref: str) -> str:
-    """Map child ref ``concept.title`` through parent ref ``song.concept``."""
-    suffix = _child_suffix(child_input_name, child_ref)
-    return f"{parent_ref}{suffix}" if suffix else parent_ref
-
-
-def _parent_prose_for_cache_ref(parent_items_by_name: Mapping[str, dict[str, Any]], parent_cache_ref: str) -> str:
-    parent_chunk = parent_items_by_name.get(parent_cache_ref)
-    if parent_chunk is None:
-        return ""
-    return str(parent_chunk.get("prose_before", ""))
-
-
-def _cache_ref_is_declared_or_covered(child_ref: str, declared: set[str]) -> bool:
-    return any(_path_is_ancestor_or_equal(declared_ref, child_ref) for declared_ref in declared)
-
-
-def _aggregate_sub_workflow_cache_candidates_by_child(
-    candidates: list[_SubWorkflowCacheCandidate],
-) -> list[_SubWorkflowCacheGroup]:
-    """Return deterministic groups with one candidate per child cache ref.
-
-    Tie-break: lex-smallest ``(parent_node_id, parent_workflow, parent_cache_ref)``
-    wins for now. Multiple distinct origins for the same cache ref are still
-    surfaced as one child edit; token/savings estimates are guarded separately
-    so we do not present one origin as measured truth when estimates disagree.
-    """
-    by_child: dict[str, dict[str, _SubWorkflowCacheCandidate]] = {}
-    for candidate in candidates:
-        by_ref = by_child.setdefault(candidate.child_workflow, {})
-        existing = by_ref.get(candidate.child_cache_ref)
-        if existing is None:
-            by_ref[candidate.child_cache_ref] = candidate
-            continue
-        merged_origin_count = existing.origin_count + candidate.origin_count
-        chosen = existing
-        if (candidate.parent_node_id, candidate.parent_workflow, candidate.parent_cache_ref) < (
-            existing.parent_node_id,
-            existing.parent_workflow,
-            existing.parent_cache_ref,
-        ):
-            chosen = candidate
-        parent_prose_origins_differ = (
-            existing.parent_prose_origins_differ
-            or candidate.parent_prose_origins_differ
-            or existing.parent_prose != candidate.parent_prose
-        )
-        by_ref[candidate.child_cache_ref] = replace(
-            chosen,
-            parent_prose="" if parent_prose_origins_differ else existing.parent_prose,
-            parent_prose_origins_differ=parent_prose_origins_differ,
-            origin_count=merged_origin_count,
-            has_multiple_parent_origins=True,
-        )
-
-    groups: list[_SubWorkflowCacheGroup] = []
-    for child_workflow in sorted(by_child):
-        candidates_for_child = tuple(by_child[child_workflow][name] for name in sorted(by_child[child_workflow]))
-        groups.append(_SubWorkflowCacheGroup(child_workflow=child_workflow, candidates=candidates_for_child))
-    return groups
-
-
-def _resolve_value_in_workflow_memo(
-    ref: str,
-    *,
-    workflow_path: str,
-    ctx: AnalysisContext,
-) -> Any | None:
-    """Resolve ``ref`` against memo cache scoped to a specific workflow path.
-
-    Cross-workflow analog to ``AnalysisContext._resolve_from_memo`` (which
-    keys on ``self.workflow_path``). For sub-workflow boundary findings the
-    parent value lives in the parent workflow, not the root.
-    """
-    if ctx.memo_cache is None:
-        return None
-    root = TemplateResolver.extract_root_node_id(ref)
-    if not root:
-        return None
-    try:
-        latest = _latest_memo_for_freshness_check(ctx.memo_cache, root, workflow_path=workflow_path, ctx=ctx)
-    except Exception:
-        logger.debug("memo_cache freshness check failed for %s in %s", ref, workflow_path, exc_info=True)
-        return None
-    if latest is None:
-        return None
-    output, _created_at = latest
-    if not isinstance(output, dict):
-        return None
-    try:
-        resolved = TemplateResolver.resolve_template(f"${{{ref}}}", {root: output})
-    except Exception:
-        logger.debug("memo resolve failed for %s in %s", ref, workflow_path, exc_info=True)
-        return None
-    if isinstance(resolved, str) and resolved == f"${{{ref}}}":
-        return None
-    return _normalize_empty(resolved)
-
-
-def _resolve_value_in_workflow_parameters(
-    ref: str,
-    *,
-    workflow_path: str,
-    ctx: AnalysisContext,
-) -> Any | None:
-    """Resolve ``ref`` against workflow-scoped parameters (Tier 0).
-
-    Cross-workflow analog to ``AnalysisContext._resolve_from_parameters``.
-    For sub-workflow boundary findings the parent value lives in the parent
-    workflow, not necessarily the analyzed root.
-
-    Reads from ``ctx.parameters_for_workflow(workflow_path)``. For the analyzed
-    root that is the caller's current parameters; for nested children that is
-    the walker-resolved parameter dict from ``_build_parameters_by_workflow``.
-    Both are honest signals because the walker resolves parent ``inputs:``
-    expressions through the same template resolver used elsewhere in analysis.
-
-    Placement before memo mirrors ``AnalysisContext.resolve_ref_value``:
-    current parameters win over historical memo values from prior runs.
-    """
-    params = ctx.parameters_for_workflow(workflow_path)
-    root = TemplateResolver.extract_root_node_id(ref)
-    if not root:
-        return None
-    if root not in params and root in ctx.parameters:
-        params = ctx.parameters
-    if root not in params:
-        return None
-    try:
-        resolved = TemplateResolver.resolve_template(f"${{{ref}}}", {root: params[root]})
-    except Exception:
-        logger.debug("parameters resolve failed for %s in %s", ref, workflow_path, exc_info=True)
-        return None
-    if isinstance(resolved, str) and resolved == f"${{{ref}}}":
-        return None
-    return _normalize_empty(resolved)
-
-
-def _trace_node_output_for(
-    node_id: str,
-    *,
-    workflow_path: str,
-    ctx: AnalysisContext,
-    cw_result: Any,
-) -> dict[str, Any] | None:
-    """Latest event output for ``(workflow_path, node_id)`` from trace.
-
-    Last match wins (loop-recovery semantics — ``workflow_trace.py`` uses the
-    final event per node id as the canonical "current state"). Prefers
-    ``event["node_output"]`` (full namespaced dict — supports dotted paths)
-    over ``event["llm_response"]`` (literal string for ``${node.response}``
-    refs on LLM nodes that didn't write a structured output).
-    """
-    if ctx.trace is None:
-        return None
-    edges = _edge_child_paths(cw_result)
-    output: dict[str, Any] | None = None
-    for we in ctx.trace.walk(edges=edges, workflow_path=ctx.workflow_path):
-        if we.workflow_path != workflow_path or we.event.get("node_id") != node_id:
-            continue
-        event_output = we.event.get("node_output")
-        if isinstance(event_output, Mapping):
-            output = dict(event_output)
-            continue
-        llm_response = we.event.get("llm_response")
-        if isinstance(llm_response, str):
-            output = {"response": llm_response}
-    return output
-
-
-def _resolve_value_in_workflow_trace(
-    ref: str,
-    *,
-    workflow_path: str,
-    ctx: AnalysisContext,
-    cw_result: Any,
-) -> Any | None:
-    """Resolve ``ref`` against trace data scoped to ``workflow_path``.
-
-    Walker attribution: ``TraceTree.walk(edges=...)`` threads workflow_path
-    via the cross-workflow edge map for sub-workflows / batch items; we
-    filter on ``we.workflow_path == workflow_path AND we.event['node_id']
-    == root`` to find the parent's event.
-    """
-    root = TemplateResolver.extract_root_node_id(ref)
-    if not root:
-        return None
-    output = _trace_node_output_for(root, workflow_path=workflow_path, ctx=ctx, cw_result=cw_result)
-    if output is None:
-        return None
-    try:
-        resolved = TemplateResolver.resolve_template(f"${{{ref}}}", {root: output})
-    except Exception:
-        logger.debug("trace resolve failed for %s in %s", ref, workflow_path, exc_info=True)
-        return None
-    if isinstance(resolved, str) and resolved == f"${{{ref}}}":
-        return None
-    return _normalize_empty(resolved)
-
-
-def _resolve_input_at_workflow_node_invocation(
-    *,
-    parent_node_id: str,
-    parent_workflow: str,
-    child_input_name: str,
-    child_cache_ref: str,
-    ctx: AnalysisContext,
-    cw_result: Any,
-) -> Any | None:
-    """Read the resolved input value from the parent's workflow-node trace event.
-
-    Closes the input-passthrough gap left by ``_resolve_value_in_workflow_trace``:
-    when the parent passes a workflow-input value (e.g. ``${concept}`` where
-    ``concept`` is the parent's own input parameter, not a node output), the
-    by-node-id walker finds nothing because no node in the parent produces
-    ``concept``. The resolved value still exists, recorded on the parent's
-    workflow-node event under ``node_params['inputs'][child_input_name]``.
-
-    The engine populates ``node_params`` with template-resolved values prior to
-    invoking the child workflow (``runtime/engine/engine.py``: ``node.params =
-    merged_params`` after ``resolve_templates``, then ``record_trace(node.params)``).
-    Reading that mapping by ``child_input_name`` is robust against complex template
-    expressions and is unaffected by the size-trimming applied to
-    ``template_resolutions`` in long-running real-world fixtures.
-
-    Last match wins (loop-recovery semantics — mirrors ``_trace_node_output_for``).
-    Returns ``None`` if the trace is missing, the parent event has no
-    ``node_params['inputs']`` mapping, or the keyed value isn't present.
-    """
-    if ctx.trace is None:
-        return None
-    edges = _edge_child_paths(cw_result)
-    resolved_value: Any = None
-    for we in ctx.trace.walk(edges=edges, workflow_path=ctx.workflow_path):
-        if we.workflow_path != parent_workflow:
-            continue
-        if we.event.get("node_id") != parent_node_id:
-            continue
-        params = we.event.get("node_params")
-        if not isinstance(params, Mapping):
-            continue
-        inputs = params.get("inputs")
-        if not isinstance(inputs, Mapping):
-            continue
-        candidate_value = inputs.get(child_input_name)
-        if candidate_value is None:
-            continue
-        resolved_value = _resolve_child_suffix_in_value(candidate_value, child_input_name, child_cache_ref)
-    return _normalize_empty(resolved_value) if resolved_value is not None else None
-
-
-def _resolve_child_suffix_in_value(value: Any, child_input_name: str, child_cache_ref: str) -> Any | None:
-    suffix = _child_suffix(child_input_name, child_cache_ref)
-    if not suffix:
-        return value
-    synthetic_ref = f"__value{suffix}"
-    if not TemplateResolver.variable_exists(synthetic_ref, {"__value": value}):
-        return None
-    return TemplateResolver.resolve_value(synthetic_ref, {"__value": value})
-
-
-def _estimate_parent_value_tokens(
-    *,
-    parent_workflow: str,
-    parent_value_expr: str,
-    parent_node_id: str,
-    child_workflow: str,
-    child_input_name: str,
-    child_cache_ref: str,
-    model: str,
-    ctx: AnalysisContext,
-    cw_result: Any,
-) -> int | None:
-    """Tokens for the parent value flowing across a sub-workflow boundary.
-
-    Tier 0: workflow parameters (caller-provided for the analyzed root, or
-    walker-propagated values for nested children). Mirrors the "parameters win
-    over memo" convention so current inputs beat stale memo from prior runs.
-    Tier 1: memo cache (cross-workflow scoped, by ``parent_value_expr`` root).
-    Tier 2: trace by node_id — for node-output-rooted refs (e.g. ``${creative.direction}``)
-    where ``creative`` is a node id in the parent workflow.
-    Tier 3: parent workflow-node ``node_params['inputs'][child_input_name]`` —
-    closes the input-passthrough case (e.g. ``${concept}`` where ``concept`` is
-    the parent's own workflow input). Reads from the runtime invocation site
-    rather than reconstructing via the upstream node lookup.
-    Tier 4: ``None`` (honest unmeasurable — never fabricate).
-
-    Coalesce expressions (``${a ?? b}``) are not handled — too ambiguous
-    which operand sourced the value at runtime; returning None keeps the
-    rest of the projection honest.
-    """
-    ref = parent_value_expr
-    if "??" in ref:
-        return None
-    workflow_path = parent_workflow
-    value = _resolve_value_in_workflow_parameters(ref, workflow_path=workflow_path, ctx=ctx)
-    if value is None:
-        value = _resolve_value_in_workflow_memo(ref, workflow_path=workflow_path, ctx=ctx)
-    if value is None:
-        value = _resolve_value_in_workflow_trace(ref, workflow_path=workflow_path, ctx=ctx, cw_result=cw_result)
-    if value is None:
-        value = _resolve_input_at_workflow_node_invocation(
-            parent_node_id=parent_node_id,
-            parent_workflow=workflow_path,
-            child_input_name=child_input_name,
-            child_cache_ref=child_cache_ref,
-            ctx=ctx,
-            cw_result=cw_result,
-        )
-    if value is None:
-        return None
-    return estimate_tokens(model, deterministic_serialize(value))[0]
-
-
-def _workflow_basename(workflow_path: str) -> str:
-    return workflow_path.rsplit("/", 1)[-1] if "/" in workflow_path else workflow_path
-
-
-_MODEL_SWITCH_BAND = 1024
-
-
-def _project_grouped_cache_savings(
-    group: _SubWorkflowCacheGroup,
-    rows_by_node_path: dict[tuple[str | None, str], PerCallRow],
-    ctx: AnalysisContext,
-    cw_result: Any,
-) -> tuple[int, float | None, tuple[_GroupedConsumerProjection, ...], dict[str, int | None], str]:
-    """Compute per-consumer cumulative cache facts for one child workflow.
-
-    Returns ``(strictest_threshold, total_savings_usd, projections,
-    tokens_per_input, strictest_model)``. Callers classify the group case from
-    ``projections`` (per-consumer per-call check) — the threshold gate is NOT a
-    cohort-total comparison. ``total_savings_usd`` is cohort-level (already
-    accounts for call count) for honest cost framing.
-    """
-    consumer_rows = _executed_consumer_rows(group, rows_by_node_path)
-    if not consumer_rows:
-        return (0, None, (), {}, "")
-
-    strictest_threshold = max(get_min_cache_tokens(row.model) for _, row in consumer_rows)
-    strictest_model = next(
-        row.model for _, row in consumer_rows if get_min_cache_tokens(row.model) == strictest_threshold
-    )
-    tokens_per_ref = _tokens_per_group_ref(
-        group,
-        consumer_rows,
-        strictest_model=strictest_model,
-        ctx=ctx,
-        cw_result=cw_result,
-    )
-    projections, total_savings = _grouped_consumer_projections(
-        group,
-        rows_by_node_path,
-        tokens_per_ref=tokens_per_ref,
-    )
-    return (strictest_threshold, total_savings, projections, tokens_per_ref, strictest_model)
-
-
-def _executed_consumer_rows(
-    group: _SubWorkflowCacheGroup,
-    rows_by_node_path: dict[tuple[str | None, str], PerCallRow],
-) -> list[tuple[str, PerCallRow]]:
-    consumer_rows: list[tuple[str, PerCallRow]] = []
-    seen: set[str] = set()
-    for candidate in group.candidates:
-        for node_id in candidate.child_node_ids:
-            if node_id in seen:
-                continue
-            row = rows_by_node_path.get((group.child_workflow, node_id))
-            if row is None or not row.model:
-                continue
-            if row.did_not_execute_in_trace:
-                continue
-            consumer_rows.append((node_id, row))
-            seen.add(node_id)
-    return consumer_rows
-
-
-def _tokens_per_group_ref(
-    group: _SubWorkflowCacheGroup,
-    consumer_rows: list[tuple[str, PerCallRow]],
-    *,
-    strictest_model: str,
-    ctx: AnalysisContext,
-    cw_result: Any,
-) -> dict[str, int | None]:
-    tokens_per_ref: dict[str, int | None] = {}
-    for candidate in group.candidates:
-        tokens = None
-        if not candidate.has_multiple_parent_origins:
-            tokens = _tokens_from_cross_workflow_rows(candidate, consumer_rows)
-        if tokens is None and not candidate.has_multiple_parent_origins:
-            tokens = _estimate_parent_value_tokens(
-                parent_workflow=candidate.parent_workflow,
-                parent_value_expr=candidate.parent_cache_ref,
-                parent_node_id=candidate.parent_node_id,
-                child_workflow=candidate.child_workflow,
-                child_input_name=candidate.child_input_name,
-                child_cache_ref=candidate.child_cache_ref,
-                model=strictest_model,
-                ctx=ctx,
-                cw_result=cw_result,
-            )
-        tokens_per_ref[candidate.child_cache_ref] = tokens
-    return tokens_per_ref
-
-
-def _cache_refs_by_consumer(group: _SubWorkflowCacheGroup) -> dict[str, list[str]]:
-    refs_by_consumer: dict[str, list[str]] = {}
-    for candidate in group.candidates:
-        for node_id in candidate.child_node_ids:
-            refs_by_consumer.setdefault(node_id, []).append(candidate.child_cache_ref)
-    return refs_by_consumer
-
-
-def _grouped_consumer_projections(
-    group: _SubWorkflowCacheGroup,
-    rows_by_node_path: dict[tuple[str | None, str], PerCallRow],
-    *,
-    tokens_per_ref: dict[str, int | None],
-) -> tuple[tuple[_GroupedConsumerProjection, ...], float | None]:
-    projections: list[_GroupedConsumerProjection] = []
-    total_savings: float | None = 0.0
-    for consumer_node_id, cache_refs in sorted(_cache_refs_by_consumer(group).items()):
-        row = rows_by_node_path.get((group.child_workflow, consumer_node_id))
-        if row is None or not row.model or row.did_not_execute_in_trace:
-            continue
-        multiplier = invocation_count_for(row)
-        missing_tokens = any(tokens_per_ref.get(ref) is None for ref in cache_refs)
-        per_call_sum = sum(tokens_per_ref[ref] or 0 for ref in cache_refs if tokens_per_ref.get(ref) is not None)
-        savings = None if missing_tokens else _estimate_token_savings_usd(row.model, per_call_sum, multiplier)
-        if savings is None:
-            total_savings = None
-        elif total_savings is not None:
-            total_savings += savings
-        projections.append(
-            _GroupedConsumerProjection(
-                consumer_node_id=consumer_node_id,
-                model=row.model,
-                consumed_inputs=tuple(cache_refs),
-                per_call_prefix_tokens=per_call_sum,
-                threshold=get_min_cache_tokens(row.model),
-                savings_usd=savings,
-                did_not_execute_in_trace=False,
-            )
-        )
-    return (tuple(projections), total_savings)
-
-
-def _tokens_from_cross_workflow_rows(
-    candidate: _SubWorkflowCacheCandidate,
-    consumer_rows: list[tuple[str, PerCallRow]],
-) -> int | None:
-    """Reuse row-level token estimates when they already exist."""
-    for _node_id, row in consumer_rows:
-        for contribution in row.cross_workflow_inputs:
-            if not isinstance(contribution, CrossWorkflowInputContribution):
-                continue
-            contribution_ref = contribution.child_cache_ref or contribution.child_input_name
-            if contribution_ref == candidate.child_cache_ref:
-                return contribution.tokens_per_call
-    return None
-
-
-def _classify_group_case(projections: tuple[_GroupedConsumerProjection, ...]) -> str:
-    """Group case = worst per-consumer case.
-
-    Each consumer's case is decided by per-call cache-prefix size against that
-    consumer's threshold:
-
-    - ``actionable``  — per_call_prefix_tokens >= consumer.threshold
-                        (cache fires for this consumer as-declared)
-    - ``model_switch`` — fires below the consumer's current threshold but at
-                        the 1,024-tier (smallest Anthropic minimum) after a
-                        model swap
-    - ``refactor``    — below every provider tier; needs content growth
-
-    Group rolls up to the worst case so the agent gets honest advice. If ANY
-    resolved consumer can't fire (e.g., per_call_prefix below 1,024), the
-    group is ``refactor`` — switching model wouldn't help that consumer.
-    """
-    if not projections:
-        return "unmeasurable"
-    per_consumer_cases = [_projection_case(projection) for projection in projections]
-    if "refactor" in per_consumer_cases:
-        return "refactor"
-    if "model_switch" in per_consumer_cases:
-        return "model_switch"
-    return "actionable"
-
-
-def _projection_case(projection: _GroupedConsumerProjection) -> str:
-    per_call = projection.per_call_prefix_tokens
-    if per_call >= projection.threshold:
-        return "actionable"
-    if per_call >= _MODEL_SWITCH_BAND:
-        return "model_switch"
-    return "refactor"
-
-
-def _split_group_by_projection_case(
-    group: _SubWorkflowCacheGroup,
-    projections: tuple[_GroupedConsumerProjection, ...],
-) -> list[tuple[_SubWorkflowCacheGroup, tuple[_GroupedConsumerProjection, ...], str]]:
-    """Split mixed-cacheability child recommendations by consumer case.
-
-    A child workflow can have one consumer whose proposed cache prefix clears
-    the provider minimum and another consumer whose narrower prefix is below
-    every provider tier. Rendering those as one recommendation forces the body
-    to choose one representative token count and produces incoherent text like
-    "~8,504 tokens per call, below the 1,024-token minimum". Split at the
-    producer boundary so every recommendation has one coherent threshold story.
-    """
-    if not projections:
-        return [(group, projections, "unmeasurable")]
-
-    by_case: dict[str, list[_GroupedConsumerProjection]] = {}
-    for projection in projections:
-        by_case.setdefault(_projection_case(projection), []).append(projection)
-
-    if len(by_case) == 1:
-        case = next(iter(by_case))
-        return [(group, projections, case)]
-
-    result: list[tuple[_SubWorkflowCacheGroup, tuple[_GroupedConsumerProjection, ...], str]] = []
-    for case in ("actionable", "model_switch", "refactor"):
-        case_projections = tuple(by_case.get(case, ()))
-        if not case_projections:
-            continue
-        consumer_ids = frozenset(projection.consumer_node_id for projection in case_projections)
-        result.append((_group_for_consumer_nodes(group, consumer_ids), case_projections, case))
-    return result
-
-
-def _group_for_consumer_nodes(group: _SubWorkflowCacheGroup, consumer_ids: frozenset[str]) -> _SubWorkflowCacheGroup:
-    candidates: list[_SubWorkflowCacheCandidate] = []
-    for candidate in group.candidates:
-        child_node_ids = tuple(node_id for node_id in candidate.child_node_ids if node_id in consumer_ids)
-        if not child_node_ids:
-            continue
-        body_refs_by_node = {
-            node_id: tuple(refs) for node_id, refs in candidate.body_refs_by_node.items() if node_id in consumer_ids
-        }
-        candidates.append(
-            replace(
-                candidate,
-                child_count=len(child_node_ids),
-                child_node_ids=child_node_ids,
-                body_refs_by_node=body_refs_by_node,
-            )
-        )
-    return _SubWorkflowCacheGroup(child_workflow=group.child_workflow, candidates=tuple(candidates))
-
-
-def _strictest_model_and_threshold(projections: tuple[_GroupedConsumerProjection, ...]) -> tuple[str, int]:
-    if not projections:
-        return "", 0
-    strictest = max(projections, key=lambda projection: projection.threshold)
-    return strictest.model, strictest.threshold
-
-
-def _total_savings_for_case(
-    projections: tuple[_GroupedConsumerProjection, ...],
-    *,
-    case: str,
-) -> float | None:
-    if case != "actionable":
-        return None
-    total = 0.0
-    for projection in projections:
-        if projection.savings_usd is None:
-            return None
-        total += projection.savings_usd
-    return total
-
-
-def _format_grouped_body_block(
-    group: _SubWorkflowCacheGroup,
-    projections: tuple[_GroupedConsumerProjection, ...],
-    tokens_per_input: dict[str, int | None],
-    strictest_model: str,
-    strictest_threshold: int,
-    case: str,
-    cw_result: Any,
-) -> str:
-    """Render the case-specific body embedded in the diagnostic message.
-
-    All threshold comparisons use ``per_call_prefix_tokens`` — the bytes the
-    provider's cache marker sees on each call. Cohort totals (per_call × calls)
-    drive ``savings_usd`` only; they're irrelevant to whether the cache fires.
-    """
-    candidates_by_ref = {candidate.child_cache_ref: candidate for candidate in group.candidates}
-    parent_label = _workflow_basename(group.candidates[0].parent_workflow) if group.candidates else "parent workflow"
-    ref_count = len(group.candidates)
-    input_phrase = _count_phrase(ref_count, "value")
-    # Per-call prefix size of the LARGEST consumer in the group — used as the
-    # representative "what does the provider see per call" number for body text.
-    # For multi-consumer groups we still render per-consumer lines below.
-    per_call_max = max((p.per_call_prefix_tokens for p in projections), default=0)
-    lines: list[str] = []
-
-    if case == "unmeasurable":
-        return _format_unmeasurable_grouped_body(
-            group=group,
-            input_phrase=input_phrase,
-            tokens_per_input=tokens_per_input,
-            cw_result=cw_result,
-        )
-
-    if case == "refactor":
-        return _format_refactor_grouped_body(
-            group=group,
-            input_phrase=input_phrase,
-            ref_count=ref_count,
-            per_call_max=per_call_max,
-            tokens_per_input=tokens_per_input,
-            cw_result=cw_result,
-        )
-
-    threshold_clause = (
-        f"above {strictest_model}'s {strictest_threshold:,}-token cache minimum"
-        if per_call_max >= strictest_threshold
-        else f"below {strictest_model}'s {strictest_threshold:,}-token cache minimum"
-    )
-    if len(projections) > 1:
-        lines.append(
-            f"{input_phrase.capitalize()} {_flow_verb(ref_count)} in from parent {parent_label}, used by "
-            f"{_count_phrase(len(projections), 'consumer node')}."
-        )
-        _append_honest_edit_lines(
-            lines, group, tokens_per_input=tokens_per_input, cw_result=cw_result, include_cleanup=False
-        )
-        lines.append("Prompt-body templates to remove and per-consumer cache prefix:")
-        for projection in projections:
-            lines.extend(
-                _format_per_consumer_input_lines(
-                    projection=projection,
-                    candidates_by_input=candidates_by_ref,
-                    tokens_per_input=tokens_per_input,
-                    cw_result=cw_result,
-                )
-            )
-    else:
-        lines.append(
-            f"{input_phrase.capitalize()} {_flow_verb(ref_count)} in from parent {parent_label}: "
-            f"{_format_tokens_phrase(per_call_max)} per call ({threshold_clause})."
-        )
-        _append_honest_edit_lines(lines, group, tokens_per_input=tokens_per_input, cw_result=cw_result)
-
-    if case == "model_switch":
-        alternatives = anthropic_models_at_threshold(_MODEL_SWITCH_BAND)
-        lines.append(
-            f"→ Switch model: replace the `- model:` line in {_workflow_basename(group.child_workflow)} with one of:"
-        )
-        for model in alternatives:
-            suffix = " (recommended — pflow's default)" if model == "claude-sonnet-4-5" else ""
-            lines.append(f"    anthropic/{model}{suffix}")
-        lines.append(
-            "  These cache at ≥1,024 tokens. `prompt_cache:` declarations transfer unchanged. "
-            "Switching providers changes base inference cost — see `pflow guide prompt-caching`."
-        )
-        lines.append("→ Then: apply steps (1)(2)(3) above.")
-        lines.append(
-            f"→ Monitor: re-run analyze-cache when per-call content grows past {strictest_threshold:,} tokens "
-            "to enable caching at the current model."
-        )
-
-    return "\n".join(lines)
-
-
-def _format_unmeasurable_grouped_body(
-    *,
-    group: _SubWorkflowCacheGroup,
-    input_phrase: str,
-    tokens_per_input: dict[str, int | None],
-    cw_result: Any,
-) -> str:
-    lines = [
-        f"{input_phrase.capitalize()} flow in but no consumer node has a resolved model — "
-        "cannot compute the cache threshold."
-    ]
-    _append_honest_edit_lines(lines, group, tokens_per_input=tokens_per_input, cw_result=cw_result)
-    lines.append(
-        "→ Set `settings.default_model` or add `- model:` to each consumer node in "
-        f"{_workflow_basename(group.child_workflow)}, then re-run analyze-cache."
-    )
-    return "\n".join(lines)
-
-
-def _format_refactor_grouped_body(
-    *,
-    group: _SubWorkflowCacheGroup,
-    input_phrase: str,
-    ref_count: int,
-    per_call_max: int,
-    tokens_per_input: dict[str, int | None],
-    cw_result: Any,
-) -> str:
-    subject = f"One value `{group.candidates[0].child_cache_ref}`" if ref_count == 1 else input_phrase.capitalize()
-    lines = [
-        f"{subject} {_format_tokens_phrase(per_call_max)} per call, "
-        "below the smallest provider cache minimum (1,024 — Anthropic Sonnet 4.5)."
-    ]
-    _append_honest_edit_lines(lines, group, tokens_per_input=tokens_per_input, cw_result=cw_result)
-    lines.append("→ Monitor: re-run analyze-cache when per-call content grows past 1,024 tokens.")
-    lines.append(f"→ Verify: confirm {_format_tokens_phrase(per_call_max)} is the realistic per-call size.")
-    return "\n".join(lines)
-
-
-def _append_honest_edit_lines(
-    lines: list[str],
-    group: _SubWorkflowCacheGroup,
-    *,
-    tokens_per_input: dict[str, int | None],
-    cw_result: Any,
-    include_cleanup: bool = True,
-) -> None:
-    if _group_has_subpath_candidates(group):
-        lines.append(_subpath_honesty_sentence(group))
-    lines.extend(_format_exact_child_cache_edits(group))
-    if include_cleanup:
-        lines.append("Prompt-body templates to remove:")
-        lines.extend(
-            _format_single_consumer_input_lines(
-                candidates=group.candidates,
-                tokens_per_input=tokens_per_input,
-                cw_result=cw_result,
-            )
-        )
-
-
-def _group_has_subpath_candidates(group: _SubWorkflowCacheGroup) -> bool:
-    return any(candidate.child_cache_ref != candidate.child_input_name for candidate in group.candidates)
-
-
-def _subpath_honesty_sentence(group: _SubWorkflowCacheGroup) -> str:
-    roots = tuple(dict.fromkeys(candidate.child_input_name for candidate in group.candidates))
-    root_text = ", ".join(f"`{root}`" for root in roots)
-    return (
-        "Only these listed values are used by prompts. Do not cache full objects like "
-        f"{root_text} unless you intentionally want every field in that object sent to the model."
-    )
-
-
-def _format_exact_child_cache_edits(group: _SubWorkflowCacheGroup) -> list[str]:
-    lines = ["Edit child workflow:", "  Add or extend ## Cache:"]
-    lines.append("    ```cache")
-    lines.extend(f"    {line}" for line in _exact_child_cache_block_content(group).split("\n"))
-    lines.append("    ```")
-    lines.append("  Add prompt_cache entries:")
-    for node_id, refs in sorted(_cache_refs_by_consumer(group).items()):
-        ordered_refs = [
-            candidate.child_cache_ref for candidate in group.candidates if candidate.child_cache_ref in refs
-        ]
-        lines.append(f"    {node_id}: prompt_cache: [{', '.join(ordered_refs)}]")
-    return lines
-
-
-def _exact_child_cache_block_content(group: _SubWorkflowCacheGroup) -> str:
-    """Render the paste-ready child ``## Cache`` block content.
-
-    Each chunk emits the var line on its own; when a matching parent chunk
-    contributes prose, a 40-char single-line preview of that prose is rendered
-    immediately above the var line. Chunks are separated by blank lines to
-    mirror the parent's ``## Cache`` visual structure. The single-line preview
-    is intentional: it stays scannable for agents and (because
-    ``_static_excerpt`` collapses internal whitespace) survives the renderer's
-    line-by-line indenting without breaking the cache block layout.
-    """
-    parts: list[str] = []
-    for candidate in group.candidates:
-        chunk = ""
-        if candidate.parent_prose.strip() and not candidate.parent_prose_origins_differ:
-            chunk += _static_excerpt(candidate.parent_prose, limit=_PARENT_PROSE_PREVIEW_LIMIT)
-            chunk += "\n\n"
-        chunk += f"${{{candidate.child_cache_ref}}}"
-        parts.append(chunk)
-    return "\n\n".join(parts)
-
-
-def _threshold_relation(tokens: int, threshold: int) -> str:
-    return "above" if tokens >= threshold else "below"
-
-
-def _parent_origin_clause(candidate: _SubWorkflowCacheCandidate) -> str | None:
-    """Sub-line text naming the parent expression when it differs from the child input name.
-
-    Surfaces parent→child data flow inside the action body for renamed inputs.
-    Returns ``None`` for same-name passthroughs (no signal to add) and for
-    multi-ref/literal parent values where ``parent_value_expr`` is empty.
-    """
-    expr = candidate.parent_value_expr
-    if not expr or candidate.parent_cache_ref == candidate.child_cache_ref:
-        return None
-    return f"flows in from parent as `${{{candidate.parent_cache_ref}}}`"
-
-
-def _format_per_consumer_input_lines(
-    *,
-    projection: _GroupedConsumerProjection,
-    candidates_by_input: dict[str, _SubWorkflowCacheCandidate],
-    tokens_per_input: dict[str, int | None],
-    cw_result: Any,
-) -> list[str]:
-    """Render input bullets under one consumer-node heading in the multi-consumer case."""
-    lines = [
-        f"  Node `{projection.consumer_node_id}` "
-        f"({_format_tokens_phrase(projection.per_call_prefix_tokens)} per call — "
-        f"{_threshold_relation(projection.per_call_prefix_tokens, projection.threshold)} "
-        f"{projection.threshold:,}-token minimum):"
-    ]
-    for input_name in projection.consumed_inputs:
-        candidate = candidates_by_input[input_name]
-        refs = _per_input_var_refs(candidate, cw_result).get(projection.consumer_node_id, ())
-        lines.append(
-            f"    • `{input_name}` {_format_nullable_tokens(tokens_per_input.get(input_name))} — "
-            f"uses {_format_var_refs(refs, fallback=input_name)}"
-        )
-        origin = _parent_origin_clause(candidate)
-        if origin is not None:
-            lines.append(f"        {origin}")
-    return lines
-
-
-def _format_single_consumer_input_lines(
-    *,
-    candidates: tuple[_SubWorkflowCacheCandidate, ...],
-    tokens_per_input: dict[str, int | None],
-    cw_result: Any,
-) -> list[str]:
-    """Render input bullets when the group has a single consumer node."""
-    lines: list[str] = []
-    for candidate in candidates:
-        refs_by_node = _per_input_var_refs(candidate, cw_result)
-        refs = tuple(ref for node_refs in refs_by_node.values() for ref in node_refs)
-        consumer_text = ", ".join(f"`{node_id}`" for node_id in candidate.child_node_ids)
-        lines.append(
-            f"  • `{candidate.child_cache_ref}` "
-            f"{_format_nullable_tokens(tokens_per_input.get(candidate.child_cache_ref))} — "
-            f"node(s) {consumer_text} use {_format_var_refs(refs, fallback=candidate.child_cache_ref)}"
-        )
-        origin = _parent_origin_clause(candidate)
-        if origin is not None:
-            lines.append(f"      {origin}")
-    return lines
-
-
-def _count_phrase(count: int, singular: str) -> str:
-    if count == 1:
-        return f"1 {singular}"
-    if singular == "value":
-        words = {2: "Two", 3: "Three", 4: "Four", 5: "Five"}
-        return f"{words.get(count, str(count))} values"
-    return f"{count} {singular}s"
-
-
-def _flow_verb(count: int) -> str:
-    return "flows" if count == 1 else "flow"
-
-
-def _format_tokens_phrase(tokens: int) -> str:
-    return f"~{tokens:,} tokens"
-
-
-def _format_nullable_tokens(tokens: int | None) -> str:
-    return "unmeasurable" if tokens is None else f"~{tokens:,} tokens"
-
-
-def _format_var_refs(refs: Iterable[str], *, fallback: str) -> str:
-    unique = tuple(dict.fromkeys(refs))
-    if not unique:
-        unique = (fallback,)
-    return ", ".join(f"`${{{ref}}}`" for ref in unique)
-
-
-def _per_input_var_refs(candidate: _SubWorkflowCacheCandidate, cw_result: Any) -> dict[str, tuple[str, ...]]:
-    """Prompt `${var}` references that must be removed before caching the input."""
-    return {node_id: tuple(refs) for node_id, refs in candidate.body_refs_by_node.items()}
-
-
-def _grouped_inputs_context(
-    group: _SubWorkflowCacheGroup,
-    tokens_per_input: dict[str, int | None],
-    cw_result: Any,
-) -> list[dict[str, Any]]:
-    """Structured input facts for JSON consumers and per-call row notes."""
-    items: list[dict[str, Any]] = []
-    for candidate in group.candidates:
-        consumer_node_ids = list(candidate.child_node_ids)
-        items.append({
-            "child_input_name": candidate.child_input_name,
-            "child_cache_ref": candidate.child_cache_ref,
-            "parent_value_expr": candidate.parent_value_expr,
-            "parent_cache_ref": candidate.parent_cache_ref,
-            "parent_workflow": candidate.parent_workflow,
-            "parent_node_id": candidate.parent_node_id,
-            "line_in_parent": candidate.line_in_parent,
-            "tokens_estimated": tokens_per_input.get(candidate.child_cache_ref),
-            "parent_prose": candidate.parent_prose,
-            "parent_prose_origins_differ": candidate.parent_prose_origins_differ,
-            "origin_count": candidate.origin_count,
-            "has_multiple_parent_origins": candidate.has_multiple_parent_origins,
-            "consumer_node_ids": consumer_node_ids,
-            "consumer_node_ids_csv": ", ".join(f"`{node_id}`" for node_id in consumer_node_ids),
-            "template_var_refs_by_node": {
-                node_id: list(refs) for node_id, refs in _per_input_var_refs(candidate, cw_result).items()
-            },
-        })
-    return items
-
-
-def _emit_sub_workflow_cache_findings(
-    candidates: list[_SubWorkflowCacheCandidate],
-    *,
-    rows_by_node_path: dict[tuple[str | None, str], PerCallRow],
-    ctx: AnalysisContext,
-    cw_result: Any,
-    call_counts_by_node: dict[tuple[str | None, str], int],
-) -> list[Diagnostic]:
-    """Emit one grouped diagnostic per child workflow."""
-    diagnostics: list[Diagnostic] = []
-    for group in _aggregate_sub_workflow_cache_candidates_by_child(candidates):
-        group_child_node_ids = tuple(sorted({node_id for c in group.candidates for node_id in c.child_node_ids}))
-        if ctx.trace is None:
-            if len(group_child_node_ids) < 2:
-                continue
-        else:
-            total_invocations = _total_observed_invocations(
-                child_workflow=group.child_workflow,
-                child_node_ids=group_child_node_ids,
-                call_counts_by_node=call_counts_by_node,
-            )
-            child_has_trace_evidence = any(
-                (group.child_workflow, node_id) in call_counts_by_node for node_id in group_child_node_ids
-            )
-            if child_has_trace_evidence and total_invocations < 2:
-                continue
-            if not child_has_trace_evidence and len(group_child_node_ids) < 2:
-                continue
-
-        _strictest_threshold, _total_savings, projections, tokens_per_input, _strictest_model = (
-            _project_grouped_cache_savings(group, rows_by_node_path, ctx, cw_result)
-        )
-        for split_group, split_projections, case in _split_group_by_projection_case(group, projections):
-            strictest_model, strictest_threshold = _strictest_model_and_threshold(split_projections)
-            diagnostics.append(
-                make_diagnostic(
-                    "cache.sub-workflow-cache-undeclared",
-                    node_id=None,
-                    affected_workflow=split_group.child_workflow,
-                    child_workflow=split_group.child_workflow,
-                    child_workflow_basename=_workflow_basename(split_group.child_workflow),
-                    affected_input_count=len(split_group.candidates),
-                    inputs=_grouped_inputs_context(split_group, tokens_per_input, cw_result),
-                    body_block=_format_grouped_body_block(
-                        split_group,
-                        split_projections,
-                        tokens_per_input,
-                        strictest_model,
-                        strictest_threshold,
-                        case,
-                        cw_result,
-                    ),
-                    case=case,
-                    savings_usd=_total_savings_for_case(split_projections, case=case),
-                )
-            )
-    return diagnostics
-
-
-def _collect_llm_nodes_referencing_path(ir: dict[str, Any], template_path: str) -> list[str]:
-    """LLM node ids whose ``params.prompt`` references ``${template_path}``.
-
-    Source-order; each LLM node listed at most once. Uses the template-pattern
-    walker to handle both bare references (``${X}``) and dotted-path references
-    (``${X.field}``, ``${X[0]}``). For coalesce expressions (``${a ?? b}``),
-    each operand is checked independently — symmetric with
-    ``_dynamic_before_static_warnings``.
-
-    Callers compute count via ``len(...)`` when they need the count. The list
-    flows through ``_SubWorkflowCacheCandidate.child_node_ids`` so the
-    ``cache.sub-workflow-cache-undeclared`` recommendation can both name the
-    affected nodes inline and project per-node cache-read savings.
-    """
-    ids: list[str] = []
-    if not isinstance(ir, dict):
-        return ids
-    nodes = ir.get("nodes")
-    if not isinstance(nodes, list):
-        return ids
-    for node in nodes:
-        if not isinstance(node, dict) or node.get("type") != "llm":
-            continue
-        prompt = node.get("params", {}).get("prompt", "")
-        if not isinstance(prompt, str):
-            continue
-        node_id = node.get("id")
-        if not isinstance(node_id, str) or not node_id or node_id in ids:
-            continue
-        node_inputs = _node_inputs(node)
-        refs = classify_prompt_refs(prompt, batch_alias=None, node_inputs=node_inputs)
-        if any(_path_is_equal_or_descendant(operand, template_path) for ref in refs for operand in ref.operand_paths):
-            ids.append(node_id)
-    return ids
-
-
-def _items_by_name(items: tuple[dict[str, Any], ...]) -> dict[str, dict[str, Any]]:
-    return {str(item["name"]): item for item in items if isinstance(item.get("name"), str)}
-
-
-# ---------------------------------------------------------------------------
-# Trace discrepancy detection
-# ---------------------------------------------------------------------------
-
-
-def _predict_cache_keys(
-    cw_result: Any,
-    ctx: AnalysisContext,
-) -> tuple[dict[tuple[str | None, str], str], list[str]]:
-    """Predict the runtime cache_key for every LLM node, scoped per workflow.
-
-    Walks every workflow's IR independently (root + descendants from
-    ``cw_result.irs_by_workflow``) and computes the byte-identical
-    cache_key the runtime would compute via ``plan_node`` — the same
-    canonical site the engine and the dry-run planner consume. This
-    bypasses ``build_plan``'s BFS-downstream mode (which sets
-    ``cache_key=None`` for child nodes whose parent took the downstream
-    path), the source of Bug 5 in the verification report.
-
-    ``compile_workflow`` + ``create_planner_shared`` are hoisted to the
-    per-workflow loop — N LLM nodes in one workflow incur ONE compile, not
-    N. ``plan_node`` is then invoked per LLM node against the shared
-    compiled+shared scaffold.
-
-    Per-node skip reasons replace the catch-all silent-skip count that the
-    legacy implementation produced — agents see exactly which node's
-    prediction failed and why.
-
-    Returns ``(predicted_keys, notes)`` where ``predicted_keys`` keys
-    ``(workflow_path, node_id) -> cache_key``. The contract matches the
-    legacy implementation; only the production path changes.
-    """
-    notes: list[str] = []
-    if ctx.memo_cache is None:
-        notes.append(
-            "Cache fidelity check skipped: this is a first run "
-            "(no prior runs to compare cached vs uncached against). "
-            "On later runs, specific cache misfires (TTL expired, chunks skipped) will appear here."
-        )
-        return {}, notes
-
-    irs_by_workflow = getattr(cw_result, "irs_by_workflow", None) or {}
-    if not irs_by_workflow:
-        # Defensive fallback: no cross-workflow walker output. Use the root
-        # IR alone (analyzer never calls _predict_cache_keys without the root).
-        irs_by_workflow = {ctx.workflow_path: dict(ctx.workflow_ir)}
-
-    predicted_keys: dict[tuple[str | None, str], str] = {}
-    skipped_input_workflows: list[str] = []
-    for workflow_path, ir in irs_by_workflow.items():
-        _predict_one_workflow(
-            workflow_path=workflow_path,
-            ir=ir,
-            ctx=ctx,
-            predicted_keys=predicted_keys,
-            notes=notes,
-            skipped_input_workflows=skipped_input_workflows,
-        )
-    if skipped_input_workflows:
-        notes.append(_format_skipped_workflows_note(skipped_input_workflows))
-    return predicted_keys, notes
-
-
-def _format_dynamic_batches_note(batches: tuple[DynamicBatchInfo, ...]) -> str | None:
-    """Aggregate runtime-template batches into ONE Note (B-4).
-
-    Workflow-type nodes whose ``batch.items`` is a ``${...}`` template can't
-    have their per-item children enumerated statically. Pre-B-4 rendering
-    emitted ~150 chars of near-identical prose per occurrence — lyrics-
-    generator's 3 batches blew up to ~500 chars of repeated content in
-    ``## Notes``. The aggregated form lists each batch's ``node_id`` +
-    ``items_expression`` once and shares the explanatory prose.
-
-    Single-batch case keeps the original phrasing for continuity with
-    pre-B-4 baselines and the existing substring-only test
-    (``test_template_items_gap_note_uses_real_analyze_cache_cli_param_wording``).
-    """
-    if not batches:
-        return None
-    if len(batches) == 1:
-        b = batches[0]
-        return (
-            f"Workflow batch {b.node_id} in {b.parent_workflow} uses items: {b.items_expression}; sub-workflow "
-            "rows for these runtime items are not in the per-call table. The displayed cost is measured from "
-            "trace events, not estimated. Pass the resolved list as a CLI parameter, or use inline static batch "
-            "items, to enable static child enumeration."
-        )
-    listing = ", ".join(f"`{b.node_id}` (items: `{b.items_expression}`)" for b in batches)
-    return (
-        f"{len(batches)} dynamic batches not in per-call table: {listing}. Batch items are computed at runtime, "
-        "so per-item rows can't be enumerated statically. Pass items as a CLI parameter or use inline static "
-        "items to list them. (Cost shown is measured from the trace, not estimated.)"
-    )
-
-
-def _format_fidelity_skip_note(target: str, reason: str, *, applicable: bool = True) -> str:
-    """Single SSoT for the "we couldn't verify cache fidelity here" notes.
-
-    Every skip-note in the discrepancy stage describes the same shape: the
-    analyzer wanted to compare predicted cache_keys against trace evidence
-    for a specific target (workflow or workflow.node), couldn't, and falls
-    back to reporting explicit skipped-chunk events from the trace. The
-    framing is jargon-free and consistent across all
-    9 emit sites so an agent reading the Notes section never has to map
-    "Discrepancy detection: predicted-key matching" → "cache fidelity
-    check" by themselves.
-
-    ``applicable=False`` switches the prefix for cases like ``cache: false``
-    on a node — the check isn't unavailable; it doesn't apply at all.
-
-    Mutation contract: if a producer ever bypasses this helper and emits
-    the old "Discrepancy detection: ..." prefix directly, the jargon
-    returns. The wording lives in this one place to prevent that drift.
-    """
-    prefix = "Cache fidelity check skipped for" if applicable else "Cache fidelity check not applicable to"
-    return f"{prefix} {target}: {reason}. Chunk-skip detection still applies."
-
-
-def _format_skipped_workflows_note(paths: list[str]) -> str:
-    """Aggregate per-sub-workflow skip notes into one summary (L-4).
-
-    Single-workflow case keeps the per-workflow detail. Multi-workflow case
-    lists up to 5 basenames; overflow as ``+N more``. Real lyrics-generator
-    runs emit 15 of these — pre-L-4 rendering blew ~4KB of repeated prose
-    into the Notes section.
-    """
-    if len(paths) == 1:
-        return _format_fidelity_skip_note(
-            paths[0],
-            "that sub-workflow declares inputs which weren't supplied as parameters",
-        )
-    shown = [Path(p).name if p and p != "<root>" else "<root>" for p in paths[:5]]
-    suffix = "" if len(paths) <= 5 else f" + {len(paths) - 5} more"
-    return _format_fidelity_skip_note(
-        f"{len(paths)} sub-workflows ({', '.join(shown)}{suffix})",
-        "they declare inputs which weren't supplied as parameters. "
-        "Pass concrete `<input>=<value>` parameters via CLI to enable per-workflow checks",
-    )
-
-
-def _attach_predicted_cache_keys(
-    ctx: AnalysisContext,
-    cw_result: Any,
-) -> tuple[AnalysisContext, dict[tuple[str | None, str], str], list[str]]:
-    """Run cache-key prediction once and store the result on the context."""
-    predicted_cache_keys: dict[tuple[str | None, str], str] = {}
-    prediction_fidelity_notes: list[str] = []
-    try:
-        predicted_cache_keys, prediction_fidelity_notes = _predict_cache_keys(cw_result, ctx)
-    except _PREDICTION_RECOVERABLE_EXCEPTIONS as exc:
-        logger.debug("memo freshness prediction disabled: %s", exc, exc_info=True)
-        _mark_all_prediction_skipped(predicted_cache_keys, cw_result, ctx)
-    return (
-        replace(
-            ctx,
-            predicted_cache_keys=predicted_cache_keys,
-            prediction_fidelity_notes=tuple(prediction_fidelity_notes),
-        ),
-        predicted_cache_keys,
-        prediction_fidelity_notes,
-    )
-
-
-def _predict_one_workflow(
-    *,
-    workflow_path: str | None,
-    ir: Mapping[str, Any],
-    ctx: AnalysisContext,
-    predicted_keys: dict[tuple[str | None, str], str],
-    notes: list[str],
-    skipped_input_workflows: list[str],
-) -> None:
-    """Compute predictions for every LLM node in one workflow IR.
-
-    Mutates ``predicted_keys``, ``notes``, and ``skipped_input_workflows``
-    in place. Extracted from ``_predict_cache_keys`` to keep that function
-    under the cyclomatic-complexity budget; the per-workflow body has its
-    own classification branches (input-gate, no-LLM-shortcut, scaffold
-    build, per-node loop) that naturally cluster together.
-    """
-    params = ctx.parameters_for_workflow(workflow_path)
-    declared_inputs_raw = ir.get("inputs")
-    declared_inputs: Mapping[str, Any] | None = (
-        declared_inputs_raw if isinstance(declared_inputs_raw, Mapping) else None
-    )
-    # Truly cold case: walker resolved NOTHING and the workflow declares
-    # inputs. Skip the whole workflow and aggregate to a single Notes
-    # summary at the caller level (L-4). Distinct from the partial-params
-    # case below where some inputs were resolved and we want to predict
-    # what we can on a per-node basis.
-    llm_nodes = [node for node in ir.get("nodes", []) if _is_llm_node(node)]
-    if not params and declared_inputs:
-        skipped_input_workflows.append(workflow_path or "<root>")
-        _mark_prediction_skipped(predicted_keys, workflow_path, llm_nodes)
-        return
-    if not llm_nodes:
-        return
-    # Partial-params case: walker resolved some inputs but not all (e.g.,
-    # a child input that flows from an upstream sub-workflow output the
-    # walker can't reach statically). Pad the missing slots with the
-    # standard placeholder so compile succeeds; then skip prediction per
-    # node for any node whose templates or referenced cache chunks touch
-    # a padded slot — its predicted cache_key would be placeholder-
-    # tainted and never match the trace. Per-node skip is silent;
-    # skipped-chunk attribution covers real branch-absent misses on those
-    # nodes.
-    padded_params, dummied_keys = _pad_inputs_for_prediction(params, declared_inputs)
-    dummied_chunks = _dummied_cache_chunks(ir, dummied_keys)
-    scaffold = _build_predict_scaffold(ir, padded_params, ctx.memo_cache, workflow_path)
-    if scaffold is None:
-        _mark_prediction_skipped(predicted_keys, workflow_path, llm_nodes)
-        return
-    for node in llm_nodes:
-        if _node_references_any(node, dummied_keys, dummied_chunks):
-            predicted_keys[(workflow_path, str(node["id"]))] = _PREDICTION_SKIPPED
-            continue
-        cache_key, skip_reason = _predict_node_with_scaffold(node, scaffold, workflow_path)
-        if cache_key is not None:
-            predicted_keys[(workflow_path, str(node["id"]))] = cache_key
-        elif skip_reason:
-            predicted_keys[(workflow_path, str(node["id"]))] = _PREDICTION_SKIPPED
-            notes.append(skip_reason)
-
-
-def _mark_prediction_skipped(
-    predicted_keys: dict[tuple[str | None, str], str],
-    workflow_path: str | None,
-    nodes: Iterable[Mapping[str, Any]],
-) -> None:
-    for node in nodes:
-        node_id = node.get("id")
-        if isinstance(node_id, str):
-            predicted_keys[(workflow_path, node_id)] = _PREDICTION_SKIPPED
-
-
-def _mark_all_prediction_skipped(
-    predicted_keys: dict[tuple[str | None, str], str],
-    cw_result: Any,
-    ctx: AnalysisContext,
-) -> None:
-    """Mark every known LLM node as attempted-but-uncheckable after prediction outage."""
-    irs_by_workflow = getattr(cw_result, "irs_by_workflow", None) or {ctx.workflow_path: dict(ctx.workflow_ir)}
-    for workflow_path, workflow_ir in irs_by_workflow.items():
-        nodes = workflow_ir.get("nodes", []) if isinstance(workflow_ir, Mapping) else []
-        _mark_prediction_skipped(
-            predicted_keys,
-            workflow_path,
-            (node for node in nodes if isinstance(node, Mapping) and _is_llm_node(node)),
-        )
-
-
-def _pad_inputs_for_prediction(
-    known_params: Mapping[str, Any],
-    declared_inputs: Mapping[str, Any] | None,
-) -> tuple[dict[str, Any], frozenset[str]]:
-    """Merge walker-resolved params with placeholders for missing inputs.
-
-    The discrepancy stage compiles every sub-workflow to predict runtime
-    cache_keys, but the cross-workflow walker can only resolve inputs
-    whose values flow statically from the parent. For inputs that come
-    from upstream node outputs (e.g. ``${upstream.results}``) the walker
-    leaves the slot empty — and ``compile_workflow`` would reject the
-    incomplete params dict with ``SchemaValidationError`` before any
-    per-node prediction could run, emitting a misleading "workflow failed
-    to compile" Note for workflows that run fine end-to-end.
-
-    Padding the missing slots with ``"__validation_placeholder__"`` lets
-    compile succeed (same idiom ``WorkflowValidator._validate_one_child_call``
-    uses for structural validation). The returned ``dummied_keys`` set
-    lets ``_predict_one_workflow`` skip prediction for any node whose
-    templates touch a placeholder; those predictions would never match
-    the trace's real values.
-    """
-    padded: dict[str, Any] = dict(known_params)
-    if not declared_inputs:
-        return padded, frozenset()
-    dummies = generate_dummy_parameters(dict(declared_inputs))
-    dummied: set[str] = set()
-    for key, placeholder in dummies.items():
-        if key not in padded:
-            padded[key] = placeholder
-            dummied.add(key)
-    return padded, frozenset(dummied)
-
-
-def _node_references_any(
-    node: Mapping[str, Any],
-    dummied_keys: frozenset[str],
-    dummied_chunks: frozenset[str],
-) -> bool:
-    """True iff ``node``'s cache_key would be placeholder-tainted.
-
-    Used by the discrepancy stage to skip cache_key prediction for nodes
-    whose inputs depend on dummied (un-walker-resolved) workflow inputs.
-    A node is tainted if EITHER:
-
-    - it declares ``prompt_cache: [name]`` referencing a chunk whose
-      ``var`` traces back to a dummied input (``dummied_chunks``), OR
-    - any ``${var}`` ref in the node's IR has a root in ``dummied_keys``.
-
-    Conservative on coalesce: if ANY operand of ``${a ?? b}`` has a root
-    in ``dummied_keys``, returns True. The alternative (only skip when
-    ALL operands are dummied) risks producing placeholder-tainted
-    predictions when the resolver happens to pick the dummied operand
-    first.
-    """
-    if not isinstance(node, Mapping):
-        return False
-    if _node_prompt_cache_touches(node, dummied_chunks):
-        return True
-    return _node_templates_touch(node, dummied_keys)
-
-
-def _node_prompt_cache_touches(node: Mapping[str, Any], dummied_chunks: frozenset[str]) -> bool:
-    """True iff the node's ``prompt_cache:`` lists any dummied chunk name."""
-    if not dummied_chunks:
-        return False
-    prompt_cache = node.get("prompt_cache")
-    if not isinstance(prompt_cache, list):
-        return False
-    return any(isinstance(name, str) and name in dummied_chunks for name in prompt_cache)
-
-
-def _node_templates_touch(node: Mapping[str, Any], dummied_keys: frozenset[str]) -> bool:
-    """True iff any ``${var}`` ref in the node's IR has a root in ``dummied_keys``.
-
-    Walks every nested string value in the node dict — broader than
-    ``_collect_llm_nodes_referencing_path`` (which only inspects
-    ``params.prompt``) because cache_key inputs come from any templated
-    field on the node (``params``, ``inputs``, ``batch``, nested
-    code-block inputs, etc.).
-    """
-    if not dummied_keys:
-        return False
-    for text in _walk_strings(node):
-        for match in TemplateResolver.TEMPLATE_PATTERN.finditer(text):
-            for operand in TemplateResolver.split_coalesce_operands(match.group(1)):
-                root = TemplateResolver.extract_root_node_id(operand)
-                if root and root in dummied_keys:
-                    return True
-    return False
-
-
-def _dummied_cache_chunks(
-    workflow_ir: Mapping[str, Any],
-    dummied_keys: frozenset[str],
-) -> frozenset[str]:
-    """Cache chunk names whose ``var`` traces back to a dummied input.
-
-    A ``## Cache`` block declares chunks with ``var: <ref>``; when a
-    node's ``prompt_cache:`` references such a chunk, the runtime
-    resolves ``var`` against parameters to produce the chunk's content.
-    If ``var``'s root is dummied, the chunk content carries the
-    placeholder and any node consuming it via ``prompt_cache`` would
-    produce a placeholder-tainted cache_key — same outcome as a direct
-    template ref to a dummied input.
-
-    Pre-computed once per workflow so the per-node check stays O(1) in
-    the number of chunks.
-    """
-    if not dummied_keys:
-        return frozenset()
-    cache = workflow_ir.get("cache") if isinstance(workflow_ir, Mapping) else None
-    if not isinstance(cache, Mapping):
-        return frozenset()
-    items = cache.get("items")
-    if not isinstance(items, list):
-        return frozenset()
-    tainted: set[str] = set()
-    for item in items:
-        if not isinstance(item, Mapping):
-            continue
-        var = item.get("var")
-        if not isinstance(var, str):
-            continue
-        root = TemplateResolver.extract_root_node_id(var)
-        if root and root in dummied_keys:
-            name = item.get("name")
-            if isinstance(name, str):
-                tainted.add(name)
-    return frozenset(tainted)
-
-
-def _walk_strings(value: Any) -> Iterable[str]:
-    """Yield every string value reachable from ``value`` via dict/list nesting."""
-    if isinstance(value, str):
-        yield value
-    elif isinstance(value, Mapping):
-        for inner in value.values():
-            yield from _walk_strings(inner)
-    elif isinstance(value, list):
-        for inner in value:
-            yield from _walk_strings(inner)
-
-
-@dataclass(frozen=True)
-class _PredictScaffold:
-    """Per-workflow scaffold reused across all LLM nodes in one workflow."""
-
-    compiled: Any
-    shared: dict[str, Any]
-    bare_nodes_by_id: dict[str, Any]
-
-
-def _build_predict_scaffold(
-    workflow_ir: Mapping[str, Any],
-    params: Mapping[str, Any],
-    memo_cache: Any,
-    workflow_path: str | None,
-) -> _PredictScaffold | None:
-    """Compile + planner-shared once per workflow.
-
-    Returns the scaffold, or ``None`` if compile/planner setup fails. The
-    unified validator (``_run_full_validation``) runs strictly earlier in
-    ``analyze()`` and surfaces structural errors via ``blocking_errors[]``
-    /``other_blocking_errors[]``, so this catch only fires for compiler-
-    internal failures the validator missed. Those degrade silently with
-    a debug log — the agent already has the actionable structural signal
-    from the validator pass, and a misleading "workflow failed to compile
-    (SchemaValidationError)" Note (the Bug 3 symptom) is worse than no
-    Note at all.
-
-    Callers inject ``_pflow_workflow_file`` so relative ``@./file.ext``
-    refs resolve against the workflow's own directory instead of CWD —
-    matches ``WorkflowValidator._validate_one_child_call``'s pattern.
-
-    Lazy imports keep the analyzer package import-cheap (mirrors
-    ``token_estimation.py``'s LiteLLM lazy-import).
-    """
-    from pflow.execution.plan import create_planner_shared
-    from pflow.registry import Registry
-    from pflow.runtime import compile_workflow
-
-    compile_params: dict[str, Any] = dict(params)
-    if workflow_path is not None:
-        compile_params.setdefault("_pflow_workflow_file", str(workflow_path))
-    try:
-        compiled = compile_workflow(dict(workflow_ir), Registry(), dict(compile_params))
-        shared = create_planner_shared(compiled, dict(compile_params), memo_cache, workflow_path)
-    except _PREDICTION_RECOVERABLE_EXCEPTIONS as exc:
-        logger.debug(
-            "predict-stage setup failed for %s: %s",
-            workflow_path or "<root>",
-            exc,
-            exc_info=True,
-        )
-        return None
-    bare_nodes_by_id = _enumerate_compiled_bare_nodes(compiled)
-    return _PredictScaffold(compiled=compiled, shared=shared, bare_nodes_by_id=bare_nodes_by_id)
-
-
-def _predict_node_with_scaffold(
-    node: dict[str, Any],
-    scaffold: _PredictScaffold,
-    workflow_path: str | None,
-) -> tuple[str | None, str | None]:
-    """Compute the cache_key for one node against a pre-built scaffold.
-
-    Returns ``(cache_key, skip_reason)`` — same shape as ``_predict_node_cache_key``
-    but doesn't recompile the workflow. Suitable for the per-workflow loop in
-    ``_predict_cache_keys``; tests that want a self-contained per-node call
-    use ``_predict_node_cache_key`` instead (it builds its own scaffold).
-    """
-    from pflow.runtime.engine.plan_node import plan_node
-
-    node_id = str(node.get("id", "?"))
-    workflow_label = workflow_path or "<root>"
-    target = f"{workflow_label}.{node_id}"
-    config = scaffold.compiled.node_configs.get(node_id)
-    if config is None:
-        return None, _format_fidelity_skip_note(target, "node missing from the compiled workflow (parser/IR mismatch)")
-    bare_node = scaffold.bare_nodes_by_id.get(node_id)
-    if bare_node is None:
-        return None, _format_fidelity_skip_note(target, "node not reachable from the workflow's start")
-    try:
-        plan = plan_node(bare_node, config, scaffold.shared)
-    except Exception as exc:
-        logger.debug("plan_node raised for %s.%s", workflow_label, node_id, exc_info=True)
-        return None, _format_fidelity_skip_note(target, f"planner raised {type(exc).__name__} during prediction")
-
-    if plan.cache_key is not None:
-        return plan.cache_key, None
-    if plan.template_exception is not None:
-        return None, _format_fidelity_skip_note(
-            target,
-            "a template reference couldn't be resolved at analysis time (depends on a runtime value)",
-        )
-    if plan.status == "cache_disabled":
-        return None, _format_fidelity_skip_note(target, "this node has `cache: false`", applicable=False)
-    return None, _format_fidelity_skip_note(target, f"planner returned no cache key (status={plan.status})")
-
-
-def _predict_node_cache_key(
-    *,
-    node: dict[str, Any],
-    workflow_ir: Mapping[str, Any],
-    params: Mapping[str, Any],
-    memo_cache: Any,
-    workflow_path: str | None,
-) -> tuple[str | None, str | None]:
-    """Self-contained per-node prediction — builds its own scaffold.
-
-    Production callers should use ``_predict_cache_keys`` (which hoists the
-    compile + shared per workflow, applies dummy padding for partial
-    walker params, and skips nodes whose templates touch a padded slot).
-    This helper is kept for direct test callers that want a single-node
-    prediction without setting up a ``cw_result`` / ``AnalysisContext``;
-    they pass real (already-resolved) params, so no padding is needed.
-    Returns ``(None, None)`` on scaffold failure — the silent-skip
-    contract matches production.
-    """
-    scaffold = _build_predict_scaffold(workflow_ir, params, memo_cache, workflow_path)
-    if scaffold is None:
-        return None, None
-    return _predict_node_with_scaffold(node, scaffold, workflow_path)
-
-
-def _enumerate_compiled_bare_nodes(compiled: Any) -> dict[str, Any]:
-    """BFS from ``compiled.start_node`` collecting node_id → bare-node."""
-    bare_nodes_by_id: dict[str, Any] = {}
-    start = getattr(compiled, "start_node", None)
-    if start is None:
-        return bare_nodes_by_id
-    queue: list[Any] = [start]
-    while queue:
-        bare = queue.pop(0)
-        bare_id = getattr(bare, "node_id", None)
-        if not isinstance(bare_id, str) or bare_id in bare_nodes_by_id:
-            continue
-        bare_nodes_by_id[bare_id] = bare
-        successors = getattr(bare, "successors", None) or {}
-        for succ in successors.values():
-            if succ is not None:
-                queue.append(succ)
-    return bare_nodes_by_id
-
-
-def _emit_discrepancy_diagnostics(
-    *,
-    ctx: AnalysisContext,
-    cw_result: Any,
-    notes: list[str],
-) -> list[Diagnostic]:
-    trace_data = ctx.trace_data
-    if trace_data is None:
-        return []
-    workflow_path = ctx.workflow_path
-    # Trace consumer rule (runtime/CLAUDE.md): gate on the major version.
-    # Per-event guards below handle missing optional fields, so future
-    # additive minor bumps are forward-compat.
-    fv = str(trace_data.get("format_version", ""))
-    if not fv.startswith("2."):
-        return []
-
-    if ctx.predicted_cache_keys or ctx.prediction_fidelity_notes:
-        predicted_keys = dict(ctx.predicted_cache_keys)
-        predict_notes = list(ctx.prediction_fidelity_notes)
-    else:
-        try:
-            predicted_keys, predict_notes = _predict_cache_keys(cw_result, ctx)
-        except _PREDICTION_RECOVERABLE_EXCEPTIONS:
-            logger.debug("trace discrepancy prediction failed", exc_info=True)
-            predicted_keys = {}
-            _mark_all_prediction_skipped(predicted_keys, cw_result, ctx)
-            predict_notes = list(ctx.prediction_fidelity_notes)
-    notes.extend(predict_notes)
-
-    diagnostics: list[Diagnostic] = []
-    from pflow.core.trace_tree import TraceTree
-
-    edge_child_paths = _edge_child_paths(cw_result)
-    try:
-        trace_tree = TraceTree.from_dict(trace_data)
-    except ValueError:
-        return []
-    # Bug 5 fix: seed the walker with the trace's actual root workflow_path so
-    # that top-level events get attributed to whoever produced them rather than
-    # being mis-attributed to the analyzed workflow. When trace was recorded
-    # for the analyzed workflow itself, this is a no-op.
-    trace_root = trace_data.get("workflow_path") or workflow_path
-    for leaf in trace_tree.iter_llm_leaves(edges=edge_child_paths, workflow_path=trace_root):
-        node_id = leaf.event_node_id if leaf.tier == "sub_workflow_descendant" else leaf.owner_node_id
-        event = dict(leaf.event)
-        leaf_workflow_path = leaf.workflow_path or workflow_path
-        llm_call = event.get("llm_call") or {}
-        if not isinstance(llm_call, dict):
-            continue
-        if llm_call.get("is_warmup"):
-            continue
-
-        chunks_skipped = llm_call.get("cache_chunks_skipped")
-        actual_key = llm_call.get("cache_key") or event.get("cache_key")
-        predicted_key = _predicted_key_for_event(predicted_keys, workflow_path=leaf_workflow_path, node_id=node_id)
-
-        # Skip when there is no discrepancy to surface. Missing key cases are
-        # covered by prediction skip notes from ``_predict_cache_keys`` rather
-        # than by noisy observable-only discrepancy attributions.
-        if not chunks_skipped and (actual_key is None or predicted_key is None or predicted_key == actual_key):
-            continue
-
-        root_cause, summary, suggestion, extra = _attribute_root_cause(
-            chunks_skipped=chunks_skipped,
-        )
-        context_extra = dict(extra)
-        context_extra.setdefault("affected_workflow", leaf_workflow_path or "<root>")
-        diagnostics.append(
-            make_diagnostic(
-                "cache.discrepancy",
-                node_id=node_id,
-                root_cause=root_cause,
-                root_cause_summary=summary,
-                suggestion=suggestion,
-                predicted_cache_key=predicted_key,
-                actual_cache_key=actual_key,
-                **context_extra,
-            )
-        )
-    return _aggregate_and_cap_discrepancies(diagnostics, max_total=20, notes=notes)
-
-
-def _predicted_key_for_event(
-    predicted_keys: Mapping[tuple[str | None, str], str],
-    *,
-    workflow_path: str | None,
-    node_id: str,
-) -> str | None:
-    """Return predicted key for a (workflow_path, node_id) pair.
-
-    ``_predict_cache_keys`` emits tuple keys exclusively. The lookup
-    here is direct — no fallback, no implicit re-keying.
-
-    Returns ``None`` for both a missing entry AND the ``_PREDICTION_SKIPPED``
-    sentinel (prediction was attempted but intentionally skipped — sub-workflow
-    placeholder taint, missing required params). The discrepancy stage treats
-    both as "no prediction available" so the catalog ID never carries the
-    internal sentinel string in ``predicted_cache_key`` and never fabricates a
-    ``key_mismatch`` attribution against a non-prediction.
-    """
-    predicted = predicted_keys.get((workflow_path, node_id))
-    if predicted is None or predicted == _PREDICTION_SKIPPED:
-        return None
-    return predicted
-
-
-def _attribute_root_cause(*, chunks_skipped: Any) -> tuple[str, str, str, dict[str, Any]]:
-    """Attribute a discrepancy event to one of two structural causes.
-
-    Returns ``(root_cause, summary, suggestion, extra_context)``.
-
-    The caller's gate guarantees we are only invoked when there is a
-    discrepancy to attribute: ``chunks_skipped`` is truthy, or both keys are
-    set and differ. No ``unknown`` fallback is reachable in this shape.
-    """
-    if chunks_skipped:
-        skipped = str(chunks_skipped[0])
-        return (
-            "chunk_skipped",
-            f"Cache chunk {skipped!r} skipped at runtime (branch absent)",
-            (
-                f"Cache chunk `{skipped}` was skipped at runtime (branch absent); "
-                "declaration is correct but rendered subset is shorter."
-            ),
-            {"skipped_chunk": skipped},
-        )
-    return (
-        "key_mismatch",
-        "Upstream value changed between predicted run and actual run",
-        "Upstream value changed between predicted run and actual run; re-run analyze-cache to refresh the prediction.",
-        {},
-    )
-
-
-def _aggregate_and_cap_discrepancies(
-    diags: list[Diagnostic],
-    *,
-    max_total: int,
-    notes: list[str] | None = None,
-) -> list[Diagnostic]:
-    groups: dict[tuple[str | None, str], list[Diagnostic]] = {}
-    for diag in diags:
-        ctx = diag.context or {}
-        groups.setdefault((diag.node_id, str(ctx.get("root_cause", "unknown"))), []).append(diag)
-
-    aggregated: list[Diagnostic] = []
-    for group in groups.values():
-        representative = group[0]
-        merged_context = {**(representative.context or {}), "affected_invocations": len(group)}
-        # ``replace`` rather than in-place mutation: ``make_diagnostic`` may
-        # share context refs across diagnostics in the same group, so mutating
-        # ``representative.context`` would leak ``affected_invocations`` to
-        # the rest of the group (silent shared-state bug).
-        aggregated.append(replace(representative, context=merged_context))
-    aggregated.sort(key=lambda diag: -int((diag.context or {}).get("affected_invocations", 1)))
-    if notes is not None and len(aggregated) > max_total:
-        notes.append(
-            f"Discrepancies: {len(aggregated) - max_total} additional group(s) suppressed by cap "
-            f"(showing top {max_total} by frequency)."
-        )
-    return aggregated[:max_total]
-
-
-# ---------------------------------------------------------------------------
-# Confidence aggregation — STRICT per DD#34 verbatim
-# ---------------------------------------------------------------------------
-
-
-def _aggregate_confidence(
-    rows: list[PerCallRow],
-) -> tuple[str, dict[str, int]]:
-    """STRICT semantics per DD#34 line 634.
-
-    - ``all rows trace`` → ``high_from_trace``
-    - ``all rows in {trace, memo}`` → ``medium_from_memo``
-    - ``any estimator/heuristic`` → ``low_no_data``
-    """
-    sources = [row.data_source for row in rows]
-    coverage: dict[str, int] = {
-        "trace": sum(1 for s in sources if s == "trace"),
-        "memo": sum(1 for s in sources if s == "memo"),
-        "estimator": sum(1 for s in sources if s == "estimator"),
-        "heuristic": sum(1 for s in sources if s == "heuristic"),
-        "total": len(sources),
-    }
-    if not sources:
-        return "low_no_data", coverage
-    if all(src == "trace" for src in sources):
-        return "high_from_trace", coverage
-    if all(src in ("trace", "memo") for src in sources):
-        return "medium_from_memo", coverage
-    return "low_no_data", coverage
-
-
-# ---------------------------------------------------------------------------
-# Summary builder
-# ---------------------------------------------------------------------------
-
-
-def _format_workflow_run_command(workflow_path: str | None, inputs: Mapping[str, Any] | None) -> str | None:
-    """Compose a paste-ready ``pflow run`` command from a workflow file path
-    and its declared inputs.
-
-    Returns ``None`` when no runnable file path exists — inline IR
-    (``workflow_path is None``) and ``ir-hash:<md5>`` lookup keys both lack a
-    file the agent could re-run. Callers thread the result onto
-    :class:`AnalysisSummary` so renderers can surface the command on
-    unavailable-cost branches without reaching into the IR themselves.
-
-    Placeholder values (``<value>``) are intentional: greenfield analysis has
-    no resolved parameters yet, and showing defaults would imply they're
-    required. The agent fills in real values when running.
-    """
-    if workflow_path is None or workflow_path.startswith("ir-hash:"):
-        return None
-    parts = [f"pflow run {workflow_path}"]
-    for name in inputs or {}:
-        parts.append(f"{name}=<value>")
-    return " ".join(parts)
-
-
-def _build_summary(
-    rows: list[PerCallRow],
-    warnings: list[Diagnostic],
-    *,
-    ttl: str | None = None,
-    ctx: AnalysisContext | None = None,
-    edge_child_paths: dict[str, str] | None = None,
-    ir_default_model: str | None = None,
-    scope_workflow_paths: frozenset[str] | None = None,
-    trace_index: TraceExecutionIndex | None = None,
-) -> AnalysisSummary:
-    """Aggregate per-call rows + warning counts into the spec's summary block.
-
-    Atomic cost primitives (Phase 5):
-
-    - ``compute_projections(rows, ...)`` produces three independent hypothetical
-      figures — ``no_cache_hypothetical_usd``,
-      ``first_run_with_cache_hypothetical_usd``,
-      ``rerun_within_ttl_hypothetical_usd``. Each carries one meaning;
-      the renderer chooses which to show based on context.
-    - ``compute_actually_paid(rows, trace, ...)`` produces the trace-driven
-      ``actually_paid_usd`` + ``actually_paid_tier`` (``UNAVAILABLE`` for
-      greenfield).
-
-    No field is overloaded — agents reading any single primitive know
-    exactly what it means independent of greenfield/trace context.
-
-    Absolute hypothetical figures still require output token data per the
-    tri-state contract (``None`` on greenfield without memo, real after at
-    least one run, partial when some priced rows have memo data and others
-    don't). Aggregate savings figures are input-only (output cost cancels)
-    and therefore work on greenfield even when absolutes stay ``None``.
-    """
-    # Lazy-import to avoid a circular when ``cost_estimation.py`` imports
-    # ``PerCallRow`` from this module at module load time.
-    from .cost_estimation import CostTier, compute_actually_paid, compute_projections
-
-    total_nodes = len(rows)
-    # Phase 6 split: count root rows vs sub-workflow rows. Mirrors the
-    # text renderer's previous inline ``sum(...)`` comprehensions in
-    # ``_format_sub_workflow_breakdown_line`` so JSON consumers see the
-    # same numbers without recomputing. ``ctx.workflow_path is None``
-    # covers the inline-IR case where the analyzed workflow has no
-    # resolved path — matches the existing ``or`` pattern in
-    # ``render_text._format_sub_workflow_breakdown_line``.
-    root_workflow_path = ctx.workflow_path if ctx is not None else None
-    root_count = sum(1 for row in rows if root_workflow_path is None or row.workflow_path == root_workflow_path)
-    sub_workflow_count = total_nodes - root_count
-    total_invocations, dynamic_batch_count = _estimate_total_invocations(rows)
-    total_input = sum(r.input_tokens_estimated * invocation_count_for(r) for r in rows)
-    # ``cacheable_tokens_estimated`` may be ``None`` for greenfield rows
-    # without memo (Option C — projection unmeasurable). Sum only the known
-    # values; None rows contribute 0 to the aggregate (honest "we don't know"
-    # rather than fabricated 0). Top-level summary still useful — agents
-    # see the partial signal from steady-state / post-run rows.
-    total_cacheable = sum((r.cacheable_tokens_estimated or 0) * invocation_count_for(r) for r in rows)
-    total_cache_active = sum((r.cache_active.tokens_estimated or 0) * invocation_count_for(r) for r in rows)
-    total_cache_ready = sum((r.cache_ready.tokens_estimated or 0) * invocation_count_for(r) for r in rows)
-    total_cache_opportunity = sum((r.cache_opportunity.tokens_estimated or 0) * invocation_count_for(r) for r in rows)
-    unknown_ready = sum(
-        1
-        for r in rows
-        if r.cache_ready.data_source not in {_PROJECTION_NOT_APPLICABLE, _PROJECTION_UNAVAILABLE}
-        and r.cache_ready.tokens_estimated is None
-    )
-    unknown_opportunity = sum(
-        1
-        for r in rows
-        if r.cache_opportunity.data_source not in {_PROJECTION_NOT_APPLICABLE, _PROJECTION_UNAVAILABLE}
-        and r.cache_opportunity.tokens_estimated is None
-    )
-    lower_ready = sum(1 for r in rows if r.cache_ready.confidence == "lower_bound")
-    lower_opportunity = sum(1 for r in rows if r.cache_opportunity.confidence == "lower_bound")
-    trace_coverage, executed_count, unexecuted_nodes = _trace_coverage_for_rows(rows, ctx)
-    models_observed_in_trace = tuple(sorted({model for row in rows for model in row.observed_models}))
-    # Stage C.1: heterogeneous rows have ``model = ""`` AND
-    # ``model_is_heterogeneous = True``. The ``if r.model`` truthy check below
-    # already short-circuits empty strings — the explicit
-    # ``not r.model_is_heterogeneous`` clause is defense-in-depth so future
-    # contributors who change the empty-string convention don't silently leak
-    # ``${item.model}`` literals back into the aggregate.
-    static_models = {r.model for r in rows if r.model and not r.model_is_heterogeneous}
-    model_set = static_models | (set(models_observed_in_trace) if trace_coverage == "complete" else set())
-    models = sorted(model_set)
-    heterogeneous_paths = tuple(sorted(r.node_path for r in rows if r.model_is_heterogeneous))
-    cache_focused = [d for d in warnings if _is_cache_focused(d)]
-    blocking_errors = sum(1 for d in cache_focused if d.severity == Severity.ERROR)
-    warnings_count = sum(1 for d in cache_focused if d.severity == Severity.WARNING)
-    info_count = sum(1 for d in cache_focused if d.severity == Severity.INFO)
-    actionable = warnings_count + info_count
-
-    output_tokens_by_node: Mapping[tuple[str | None, str] | str, int | None] = {
-        (r.workflow_path, r.node_path): r.output_tokens_estimated for r in rows
-    }
-    projections = compute_projections(rows, output_tokens_by_node=output_tokens_by_node, ttl=ttl)
-    actually_paid = compute_actually_paid(
-        rows,
-        trace=ctx.trace if ctx is not None else None,
-        edges=edge_child_paths,
-        scope_workflow_paths=scope_workflow_paths,
-    )
-
-    # Partial flag: actually-paid trace_partial OR projection partial. The
-    # renderer uses one boolean to decide whether to mark numbers as
-    # incomplete (which can happen on EITHER stream independently).
-    partial_cost_usd = actually_paid.tier == CostTier.TRACE_PARTIAL or projections.partial
-
-    no_cache_baseline = "no_cache_hypothetical_usd"
-    first_run_delta = _cost_delta(
-        baseline_value=projections.no_cache_hypothetical_usd,
-        compared_value=projections.first_run_with_cache_hypothetical_usd,
-        baseline=no_cache_baseline,
-        compared_to="first_run_with_cache_hypothetical_usd",
-    )
-    rerun_delta = _cost_delta(
-        baseline_value=projections.no_cache_hypothetical_usd,
-        compared_value=projections.rerun_within_ttl_hypothetical_usd,
-        baseline=no_cache_baseline,
-        compared_to="rerun_within_ttl_hypothetical_usd",
-    )
-    if trace_coverage == "none":
-        actual_vs_no_cache_delta = _unavailable_delta(
-            no_cache_baseline,
-            "actually_paid_usd",
-            reason="no_trace",
-        )
-    elif projections.absolute_exclusions:
-        # Projection cohort excludes some rows (heterogeneous batch, unpriced
-        # model, etc.). Compute savings on the priced subset when total paid
-        # is known and the projection cohort has at least one priced row to
-        # compare against; otherwise the comparison stays genuinely
-        # unavailable (renderer's elif branch surfaces the exclusion list).
-        # Each excluded row's cost was already attributed to actually_paid
-        # under the same coercion rule as ``row.cost_usd`` (trace_tree treats
-        # unpriced events as 0.0 in both totals), so the subtraction is
-        # internally consistent. Per-exclusion None values are coerced to 0.0.
-        excluded = projections.absolute_exclusions
-        no_cache = projections.no_cache_hypothetical_usd
-        if actually_paid.total_usd is not None and no_cache is not None and no_cache > 0:
-            excluded_total = sum((e.actual_cost_usd or 0.0) for e in excluded)
-            actually_paid_subset = actually_paid.total_usd - excluded_total
-            excluded_paths = tuple(
-                e.node_path
-                for e in sorted(
-                    excluded,
-                    key=lambda x: (x.workflow_path or "", x.node_path),
-                )
-            )
-            actual_vs_no_cache_delta = _cost_delta(
-                baseline_value=no_cache,
-                compared_value=actually_paid_subset,
-                baseline=no_cache_baseline,
-                compared_to="actually_paid_priced_cohort_usd",
-                excluded_nodes=excluded_paths,
-            )
-        else:
-            actual_vs_no_cache_delta = _unavailable_delta(
-                no_cache_baseline,
-                "actually_paid_usd",
-                reason="projection_exclusions",
-            )
-    else:
-        actual_vs_no_cache_delta = _cost_delta(
-            baseline_value=projections.no_cache_hypothetical_usd,
-            compared_value=actually_paid.total_usd,
-            baseline=no_cache_baseline,
-            compared_to="actually_paid_usd",
-        )
-
-    # Trace transparency fields: only populated when a trace was actually
-    # loaded (autoload or explicit). ``trace_data is None`` → both fields
-    # stay ``None`` (renderer reads ``analysis.trace_path`` to decide
-    # whether to show the header line). Back-compat fallback mirrors
-    # ``_trace_coverage_for_rows``: absent ``final_status`` → ``"success"``.
-    if ctx is not None and ctx.trace_data is not None:
-        trace_final_status: str | None = str(ctx.trace_data.get("final_status") or "success")
-        recorded = ctx.trace_data.get("start_time")
-        trace_recorded_at: str | None = str(recorded) if recorded is not None else None
-    else:
-        trace_final_status = None
-        trace_recorded_at = None
-
-    return AnalysisSummary(
-        actually_paid_usd=actually_paid.total_usd,
-        actually_paid_tier=actually_paid.tier,
-        no_cache_hypothetical_usd=projections.no_cache_hypothetical_usd,
-        first_run_with_cache_hypothetical_usd=projections.first_run_with_cache_hypothetical_usd,
-        rerun_within_ttl_hypothetical_usd=projections.rerun_within_ttl_hypothetical_usd,
-        first_run_delta=first_run_delta,
-        rerun_delta=rerun_delta,
-        actual_vs_no_cache_delta=actual_vs_no_cache_delta,
-        trace_coverage=trace_coverage,
-        evidence_scope=_evidence_scope_for_trace_coverage(trace_coverage),
-        trace_llm_nodes_static=total_nodes,
-        trace_llm_nodes_executed=executed_count,
-        trace_unexecuted_llm_rows=unexecuted_nodes,
-        blocking_errors=blocking_errors,
-        actionable_opportunities=actionable,
-        warnings_count=warnings_count,
-        info_count=info_count,
-        total_llm_nodes_estimated=total_nodes,
-        total_llm_invocations_estimated=total_invocations,
-        dynamic_batch_node_count=dynamic_batch_count,
-        total_input_tokens_estimated=total_input,
-        total_cacheable_tokens_estimated=total_cacheable,
-        total_cache_active_tokens_estimated=total_cache_active,
-        total_cache_ready_tokens_estimated=total_cache_ready,
-        total_cache_opportunity_tokens_estimated=total_cache_opportunity,
-        total_cache_ready_confidence=_aggregate_projection_confidence(row.cache_ready for row in rows),
-        total_cache_opportunity_confidence=_aggregate_projection_confidence(row.cache_opportunity for row in rows),
-        unknown_cache_ready_row_count=unknown_ready,
-        unknown_cache_opportunity_row_count=unknown_opportunity,
-        lower_bound_cache_ready_row_count=lower_ready,
-        lower_bound_cache_opportunity_row_count=lower_opportunity,
-        models_in_use=tuple(models),
-        observed_models_in_trace=models_observed_in_trace,
-        ir_default_model=ir_default_model,
-        partial_cost_usd=partial_cost_usd,
-        unavailable_models=projections.unavailable_models,
-        unavailable_models_by_workflow=_unavailable_models_by_workflow(rows),
-        heterogeneous_model_node_count=len(heterogeneous_paths),
-        heterogeneous_model_node_paths=heterogeneous_paths,
-        root_llm_node_count=root_count,
-        sub_workflow_llm_node_count=sub_workflow_count,
-        projection_exclusions=projections.absolute_exclusions,
-        trace_final_status=trace_final_status,
-        trace_recorded_at=trace_recorded_at,
-        trace_provider_llm_call_count=trace_index.provider_llm_call_count if trace_index is not None else 0,
-        trace_local_memo_llm_hit_count=trace_index.local_memo_llm_hit_count if trace_index is not None else 0,
-        trace_local_in_process_llm_hit_count=(
-            trace_index.local_in_process_llm_hit_count if trace_index is not None else 0
-        ),
-        trace_local_cache_input_tokens=trace_index.local_cache_input_tokens if trace_index is not None else 0,
-        trace_provider_cache_creation_input_tokens=(
-            trace_index.provider_cache_creation_input_tokens if trace_index is not None else 0
-        ),
-        trace_provider_cache_read_input_tokens=(
-            trace_index.provider_cache_read_input_tokens if trace_index is not None else 0
-        ),
-        stale_memo_skipped_count=len(ctx.stale_memo_skipped) if ctx is not None else 0,
-        stale_memo_uncheckable_count=len(ctx.stale_memo_uncheckable) if ctx is not None else 0,
-    )
-
-
-def _trace_coverage_for_rows(
-    rows: list[PerCallRow],
-    ctx: AnalysisContext | None,
-) -> tuple[str, int, tuple[TraceUnexecutedLLMRow, ...]]:
-    """Classify trace coverage over the static LLM rows.
-
-    Returns ``"truncated"`` only when the trace's ``final_status`` is
-    ``"failed"`` AND some static rows didn't execute (workflow died mid-run,
-    cost-projection cohort genuinely incomplete). Returns ``"complete"``
-    otherwise — including the case where ``final_status`` is success but
-    some rows didn't execute, which is normal conditional dispatch
-    (a router routes inputs to one of N branches; only one fires).
-
-    The ``final_status`` field is written by
-    ``runtime/workflow_trace.py::WorkflowTraceCollector`` from per-node
-    final-event success/failure (see ``_determine_trace_status``).
-    Defensively defaults to ``"success"`` when missing (legacy 2.0.0
-    fixtures, hand-built test traces).
-    """
-    if ctx is None or ctx.trace_data is None:
-        return "none", 0, ()
-    unexecuted = tuple(
-        TraceUnexecutedLLMRow(workflow_path=row.workflow_path, node_path=row.node_path)
-        for row in sorted(
-            (row for row in rows if row.did_not_execute_in_trace),
-            key=lambda row: (row.workflow_path or "", row.node_path),
-        )
-    )
-    executed = len(rows) - len(unexecuted)
-    if not rows:
-        return "complete", 0, ()
-    final_status = str(ctx.trace_data.get("final_status") or "success")
-    if unexecuted and final_status == "failed":
-        return "truncated", executed, unexecuted
-    return "complete", executed, unexecuted
-
-
-def _evidence_scope_for_trace_coverage(trace_coverage: str) -> str:
-    if trace_coverage == "truncated":
-        return "truncated_trace_executed_subset"
-    if trace_coverage == "complete":
-        return "complete_trace"
-    return "static_analysis"
-
-
-def _aggregate_projection_confidence(projections: Iterable[CacheProjection]) -> str:
-    values = {
-        projection.confidence
-        for projection in projections
-        if projection.data_source not in {_PROJECTION_NOT_APPLICABLE, _PROJECTION_UNAVAILABLE}
-    }
-    if not values or "unknown" in values:
-        return "unknown"
-    if "lower_bound" in values:
-        return "lower_bound"
-    if values == {"observed"}:
-        return "observed"
-    return "exact"
-
-
-def _filter_trace_dependent_warnings(warnings: list[Diagnostic]) -> list[Diagnostic]:
-    """Drop diagnostics whose catalog spec opts in via ``requires_complete_trace``.
-
-    Applied only when ``trace_coverage == "truncated"`` (workflow died mid-run,
-    so cost-projection cohorts are misleading). IR-derived findings (the
-    default) flow regardless because they describe workflow structure, not
-    execution evidence — the contract documented at
-    ``prompt_cache_analysis/CLAUDE.md`` § "Trace coverage is first-class".
-
-    Lookup mirrors ``resolve_headline_for`` (warning_catalog.py:1236-1238) —
-    catalog SSoT consulted by ID at runtime.
-    """
-    return [
-        warning
-        for warning in warnings
-        if not (
-            warning.id
-            and warning.id in CACHE_WARNING_CATALOG
-            and CACHE_WARNING_CATALOG[warning.id].requires_complete_trace
-        )
-    ]
-
-
-def _cost_delta(
-    *,
-    baseline_value: float | None,
-    compared_value: float | None,
-    baseline: str,
-    compared_to: str,
-    excluded_nodes: tuple[str, ...] = (),
-) -> CostDelta:
-    if baseline_value is None or compared_value is None or baseline_value <= 0:
-        return _unavailable_delta(baseline, compared_to)
-    delta = baseline_value - compared_value
-    if abs(delta) < 0.0000001:
-        return CostDelta(
-            amount_usd=0.0,
-            pct_of_baseline=0,
-            kind="break_even",
-            baseline=baseline,
-            compared_to=compared_to,
-            excluded_nodes=excluded_nodes,
-        )
-    return CostDelta(
-        amount_usd=abs(delta),
-        pct_of_baseline=round(100 * abs(delta) / baseline_value),
-        kind="savings" if delta > 0 else "cost_increase",
-        baseline=baseline,
-        compared_to=compared_to,
-        excluded_nodes=excluded_nodes,
-    )
-
-
-def _unavailable_delta(baseline: str, compared_to: str, *, reason: str | None = None) -> CostDelta:
-    return CostDelta(
-        amount_usd=None,
-        pct_of_baseline=None,
-        kind="unavailable",
-        baseline=baseline,
-        compared_to=compared_to,
-        unavailable_reason=reason,
-    )
-
-
-def _estimate_total_invocations(rows: list[PerCallRow]) -> tuple[int | None, int]:
-    """Estimate runtime LLM invocations from static sizes or observed trace counts."""
-    total = 0
-    dynamic_batch_count = 0
-    for row in rows:
-        if row.is_batch and row.batch_size_estimated is None and row.observed_call_count <= 0:
-            dynamic_batch_count += 1
-        else:
-            total += invocation_count_for(row)
-    if dynamic_batch_count:
-        return None, dynamic_batch_count
-    return total, dynamic_batch_count
-
-
-def _unavailable_models_by_workflow(rows: list[PerCallRow]) -> dict[str | None, tuple[str, ...]]:
-    """Return unpriced models grouped by workflow path for JSON/text attribution."""
-    from .cost_estimation import get_model_pricing
-
-    grouped: dict[str | None, set[str]] = {}
-    for row in rows:
-        if row.did_not_execute_in_trace or row.model_is_heterogeneous or not row.model:
-            continue
-        if get_model_pricing(row.model) is None:
-            grouped.setdefault(row.workflow_path, set()).add(row.model)
-    return {workflow_path: tuple(sorted(models)) for workflow_path, models in grouped.items()}
-
-
-def _safe_pct_or_none(numerator: float | None, denominator: float | None) -> int | None:
-    """Compute a percent only when both numerator and denominator are real."""
-    if numerator is None or denominator is None or denominator <= 0:
-        return None
-    return round(100 * numerator / denominator)
-
-
-# ---------------------------------------------------------------------------
-# Gemini telemetry note (Spike 1 outcome — last in note ordering)
-# ---------------------------------------------------------------------------
-
-
-_GEMINI_TELEMETRY_NOTE = (
-    "Gemini caching: Gemini reports cache reads via 'cache_read_input_tokens' "
-    "in trace events. Cache writes don't show in telemetry, so verify caching "
-    "is working by checking that follow-up calls show cache reads. Explicit "
-    "## Cache declarations are required for Gemini."
-)
-
-
-def _maybe_append_gemini_note(rows: list[PerCallRow], notes: list[str]) -> None:
-    """Append the Gemini telemetry note if any analyzed call targets Gemini."""
-    for row in rows:
-        if not row.model:
-            continue
-        try:
-            provider_info = detect_provider(row.model)
-        except Exception:
-            # detect_provider raises on malformed model strings; the row is
-            # still analyzable — skip the gemini note for it but surface at
-            # debug so a typo isn't completely silent.
-            logger.debug("detect_provider failed for model %r", row.model, exc_info=True)
-            continue
-        if provider_info is None:
-            continue
-        if provider_info.name == "gemini":
-            notes.append(_GEMINI_TELEMETRY_NOTE)
-            return
 
 
 __all__ = [
