+++ src/pflow/core/prompt_cache_analysis/stages/discrepancy/__init__.py
@@ -1,0 +1,10 @@
+"""Discrepancy prediction and diagnosis stage."""
+
+from .diagnose import _emit_discrepancy_diagnostics
+from .predict import _attach_predicted_cache_keys, _format_dynamic_batches_note
+
+__all__ = [
+    "_attach_predicted_cache_keys",
+    "_emit_discrepancy_diagnostics",
+    "_format_dynamic_batches_note",
+]
