+++ tests/test_core/test_sub_workflow_resolver.py
@@ -310,7 +310,7 @@ def test_collect_llm_nodes_referencing_path_handles_inputs_indirection(
     expected_found: bool,
 ) -> None:
     """Mutation contract: removing classifier dealiasing misses indirect cases."""
-    from pflow.core.prompt_cache_analysis.analyze import _collect_llm_nodes_referencing_path
+    from pflow.core.prompt_cache_analysis.stages.cross_workflow import _collect_llm_nodes_referencing_path
 
     params: dict[str, object] = {"prompt": prompt}
     if node_inputs is not None:
@@ -341,8 +341,8 @@ def test_resolve_sub_workflow_cross_workflow_walker_sees_resolved_prompts(tmp_pa
     real-world workflow that follows the documented external-prompt-file
     pattern.
     """
-    from pflow.core.prompt_cache_analysis.analyze import _collect_llm_nodes_referencing_path
     from pflow.core.prompt_cache_analysis.cross_workflow import walk_cross_workflow
+    from pflow.core.prompt_cache_analysis.stages.cross_workflow import _collect_llm_nodes_referencing_path
     from pflow.execution.workflow_resolver import resolve_workflow
 
     # Child workflow with external prompt that template-references shared_value.
